feat(ui-adjustements): add page adjustments and responsiveness
