+++ package.json
@@ -66,6 +66,7 @@
     "react-icons": "^5.5.0",
     "react-redux": "^9.2.0",
     "react-resizable-panels": "^2.1.7",
+    "react-responsive": "^10.0.1",
     "react-select": "^5.10.1",
     "react-spinners": "^0.15.0",
     "recharts": "^2.15.1",
