--- src/app/(app)/_components/Sittingperson.tsx
@@ -1,287 +1,0 @@
-import React from "react";
-
-const Sittingperson = ({ className }: { className: string }) => {
-  return (
-    <div className={className}>
-      <svg
-        viewBox="0 0 144 176"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M46.8253 143.548C45.3452 142.63 45.1337 137.674 51.2658 136.756C55.0719 136.206 65.6445 139.51 70.5078 141.713C70.9307 143.548 71.1422 145.2 69.8735 146.669C68.3933 148.137 65.8559 148.321 63.7414 148.137C59.7238 147.77 49.3627 145.2 46.8253 143.548Z"
-          fill="#D58F77"
-        />
-        <path
-          d="M76.4286 139.51C74.5256 138.041 75.3714 132.901 80.2348 132.718C83.8294 132.718 89.3272 138.775 93.7676 141.162C93.9791 142.814 93.9791 144.466 92.4989 145.751C91.0188 147.036 88.6928 146.852 86.5783 146.485C82.9836 145.934 78.5432 141.345 76.4286 139.51Z"
-          fill="#D58F77"
-        />
-        <path
-          d="M47.0366 142.079C45.5564 141.162 45.1335 137.123 51.2656 136.389C55.0717 135.838 65.6443 138.224 70.5076 140.427C70.9305 142.263 71.142 143.915 69.8733 145.384C68.3931 146.852 65.8557 147.036 63.7412 146.852C59.7236 146.118 49.574 143.731 47.0366 142.079Z"
-          fill="#263045"
-        />
-        <path
-          d="M76.4282 137.674C75.1595 136.573 75.1595 132.901 81.0801 132.718C84.6748 132.534 89.5382 137.123 93.7672 139.51C93.9787 141.162 93.9787 142.814 92.4985 144.099C91.0183 145.384 88.6924 145.2 86.5779 144.833C82.9832 144.099 78.5427 139.51 76.4282 137.674Z"
-          fill="#263045"
-        />
-        <path
-          d="M71.3538 110.506C68.8163 111.424 66.9133 113.26 66.4904 115.463C63.5301 125.926 59.301 138.959 59.301 140.978C59.0896 142.814 62.0499 144.282 63.953 144.65C67.9705 145.384 69.0278 144.833 70.2965 143.548C70.9309 142.814 80.6576 122.622 83.8294 105.55C79.8118 108.12 75.5828 109.038 71.3538 110.506Z"
-          fill="#1E2940"
-        />
-        <path
-          d="M83.6179 103.347C84.0408 119.685 82.3492 138.959 82.9836 140.978C84.2523 143.732 92.9218 144.282 93.5561 141.896C93.7676 140.795 97.7852 120.419 97.5737 98.9417C92.7103 101.328 88.2699 100.41 83.6179 103.347Z"
-          fill="#1F2A40"
-        />
-        <path
-          d="M130.137 125.191L83.8293 101.879L83.6178 45.8911L129.926 69.2039L130.137 125.191Z"
-          fill="#C6C6C6"
-        />
-        <path
-          d="M47.2485 105.183L0.940674 81.8701L90.173 37.2637L136.481 60.3929L47.2485 105.183Z"
-          fill="white"
-        />
-        <path
-          d="M47.2485 166.494L1.15212 143.365L0.940674 81.8704L47.2485 105.183V166.494Z"
-          fill="#EBECF3"
-        />
-        <path
-          d="M136.481 60.3931V121.887L130.137 125.192L129.926 69.2042L53.5918 107.386L53.8033 163.373L47.2483 166.494V105.183L136.481 60.3931Z"
-          fill="#DDDDE1"
-        />
-        <path
-          d="M107.089 66.8179L87.2123 56.9054V54.519L107.089 64.6151V66.8179Z"
-          fill="#3A4B8F"
-        />
-        <path
-          d="M107.609 63.5355L87.7325 53.623L102.225 46.2585L122.102 56.1711L107.609 63.5355Z"
-          fill="#5971CF"
-        />
-        <path
-          d="M107.089 64.6152L87.2123 54.7027C87.2123 54.3356 87.6352 53.7849 88.0581 53.6013L107.935 63.5138C107.3 63.6974 107.089 64.0645 107.089 64.6152Z"
-          fill="#4C61B4"
-        />
-        <path
-          d="M107.723 67.1848C107.3 67.3684 106.877 67.1848 106.877 66.8177V64.4313C106.877 64.0642 107.3 63.5135 107.723 63.33L122.101 56.1709V56.538L121.679 56.7216L120.833 58.3737L121.679 59.6586L122.101 59.8422V60.0258L107.723 67.1848Z"
-          fill="#4F63A3"
-        />
-        <path
-          d="M107.3 66.8178C107.3 67.0013 107.3 67.0013 107.3 67.0013H107.511L121.89 59.8423L121.467 59.6587L120.621 58.5573L121.467 56.7217L107.723 63.6972C107.511 63.6972 107.3 63.8807 107.3 64.0643C107.088 64.2479 107.088 64.4314 107.088 64.615L107.3 66.8178Z"
-          fill="#5E71BA"
-        />
-        <path
-          d="M121.678 56.7217L107.934 63.6972C107.723 63.6972 107.511 63.8807 107.511 64.0643C107.3 64.2479 107.3 64.4314 107.3 64.615V66.8178L121.678 59.6587V56.7217Z"
-          fill="#FFFFFA"
-        />
-        <path
-          d="M107.511 62.5958L87.635 52.4997V50.2969L107.511 60.2094V62.5958Z"
-          fill="#1935CA"
-        />
-        <path
-          d="M108.358 59.1082L88.4813 49.1956L102.649 41.853L122.525 51.9491L108.358 59.1082Z"
-          fill="#1935CA"
-        />
-        <path
-          d="M107.511 60.2097L87.635 50.2972C87.635 49.9301 88.0579 49.3794 88.4808 49.1958L108.357 59.1083C107.934 59.2919 107.511 59.8426 107.511 60.2097Z"
-          fill="#1935CA"
-        />
-        <path
-          d="M108.358 62.9631C107.935 63.1467 107.512 62.9631 107.512 62.596V60.2097C107.512 59.8425 107.935 59.2918 108.358 59.1083L122.736 51.9492V52.3164L122.313 52.4999L121.468 54.152L122.313 55.437L122.736 55.6205V55.8041L108.358 62.9631Z"
-          fill="#1935CA"
-        />
-        <path
-          d="M107.723 62.5959C107.723 62.7794 107.723 62.7794 107.723 62.7794H107.935L122.313 55.6204L121.891 55.4368L121.045 54.3354L121.891 52.4998L108.146 59.4752C107.935 59.4752 107.723 59.6588 107.723 59.8424C107.512 60.0259 107.512 60.2095 107.512 60.3931L107.723 62.5959Z"
-          fill="#1935CA"
-        />
-        <path
-          d="M122.101 52.3164L108.357 59.2919C108.146 59.2919 107.934 59.4755 107.934 59.659C107.723 59.8426 107.723 60.0262 107.723 60.2097V62.4125L122.101 55.2535V52.3164Z"
-          fill="#FFFFFA"
-        />
-        <path
-          d="M97.1504 48.6449L102.014 46.2585L101.802 46.075L96.939 48.4613L97.1504 48.6449Z"
-          fill="#E9BFE1"
-        />
-        <path
-          d="M97.1507 50.1131L104.763 46.2582L104.34 45.8911L96.7278 49.746L97.1507 50.1131Z"
-          fill="#E9BFE1"
-        />
-        <path
-          d="M98.4194 50.6639L106.032 46.8091L105.609 46.6255L97.9965 50.4804L98.4194 50.6639Z"
-          fill="#E9BFE1"
-        />
-        <path
-          d="M99.2654 69.0207L89.7501 64.248C89.5387 64.0645 89.1158 64.0645 88.9043 64.248L59.3012 79.1168C59.0897 79.1168 59.0897 79.3004 59.0897 79.3004V80.4018C59.0897 80.4018 59.0897 80.5854 59.3012 80.5854L68.8165 85.3581C69.0279 85.5416 69.4508 85.5416 69.6623 85.3581L99.2654 70.4893C99.4769 70.4893 99.4769 70.3057 99.4769 70.3057V69.3879C99.2654 69.0207 99.2654 69.0207 99.2654 69.0207Z"
-          fill="#172030"
-        />
-        <path
-          d="M68.6051 74.3442L59.0898 79.1169C58.8783 79.1169 58.8783 79.3005 58.8783 79.3005V80.4019C58.8783 80.4019 58.8783 80.5855 59.0898 80.5855L68.6051 85.3582V74.3442Z"
-          fill="#172030"
-        />
-        <path
-          d="M99.2653 69.3878L69.6621 84.2566C69.4507 84.4402 69.0278 84.4402 68.8163 84.2566L59.301 79.4839C59.0896 79.3004 59.0896 79.1168 59.301 78.9332L88.9042 64.0644C89.1156 63.8809 89.5385 63.8809 89.75 64.0644L99.2653 68.8371C99.4767 69.0207 99.4767 69.2043 99.2653 69.3878Z"
-          fill="#1E2A3D"
-        />
-        <path
-          opacity="0.9"
-          d="M63.1074 79.4841L61.8387 80.0348C61.6273 80.0348 61.4158 80.0348 61.4158 80.0348L60.1471 79.4841C59.9357 79.4841 59.9357 79.3006 60.1471 79.3006L61.4158 78.7499C61.6273 78.7499 61.8387 78.7499 61.8387 78.7499L62.896 79.3006C63.1074 79.3006 63.1074 79.3006 63.1074 79.4841ZM65.222 78.1992L64.1647 77.6485C63.9532 77.6485 63.7418 77.6485 63.7418 77.6485L62.2616 78.1992C62.0502 78.1992 62.0502 78.3828 62.2616 78.3828L63.3189 78.9335C63.5303 78.9335 63.7418 78.9335 63.7418 78.9335L65.0105 78.3828C65.222 78.3828 65.222 78.1992 65.222 78.1992ZM67.125 77.0978L66.0678 76.5471C65.8563 76.5471 65.6449 76.5471 65.6449 76.5471L64.3762 77.2814C64.1647 77.2814 64.1647 77.4649 64.3762 77.4649L65.4334 78.0156C65.6449 78.0156 65.8563 78.0156 65.8563 78.0156L67.125 77.4649C67.3365 77.2814 67.3365 77.0978 67.125 77.0978ZM69.2395 76.18L68.1823 75.6293C67.9708 75.6293 67.7594 75.6293 67.7594 75.6293L66.4907 76.18C66.2792 76.18 66.2792 76.3635 66.4907 76.3635L67.5479 76.9142C67.7594 76.9142 67.9708 76.9142 67.9708 76.9142L69.2395 76.3635C69.451 76.18 69.451 76.18 69.2395 76.18ZM71.354 75.0786L70.2968 74.5279C70.0853 74.5279 69.8739 74.5279 69.8739 74.5279L68.6052 75.0786C68.3937 75.0786 68.3937 75.2621 68.6052 75.2621L69.6624 75.8128C69.8739 75.8128 70.0853 75.8128 70.0853 75.8128L71.354 75.2621V75.0786ZM73.4685 73.9772L72.4113 73.4265C72.1998 73.4265 71.9884 73.4265 71.9884 73.4265L70.7197 73.9772C70.5082 73.9772 70.5082 74.1608 70.7197 74.1608L71.7769 74.7115C71.9884 74.7115 72.1998 74.7115 72.1998 74.7115L73.4685 74.1608C73.4685 74.1608 73.4685 74.1608 73.4685 73.9772ZM75.3716 73.0594L74.3144 72.5087C74.1029 72.5087 73.8914 72.5087 73.8914 72.5087L72.6227 73.0594C72.4113 73.0594 72.4113 73.2429 72.6227 73.2429L73.68 73.7936C73.8914 73.7936 74.1029 73.7936 74.1029 73.7936L75.3716 73.2429C75.5831 73.0594 75.5831 73.0594 75.3716 73.0594ZM77.4861 71.958L76.4289 71.4073C76.2174 71.4073 76.006 71.4073 76.006 71.4073L74.7373 71.958C74.5258 71.958 74.5258 72.1415 74.7373 72.1415L75.7945 72.6922C76.006 72.6922 76.2174 72.6922 76.2174 72.6922L77.4861 72.1415C77.6976 72.1415 77.6976 71.958 77.4861 71.958ZM79.6006 70.8566L78.5434 70.3059C78.3319 70.3059 78.1205 70.3059 78.1205 70.3059L76.8518 70.8566C76.6403 70.8566 76.6403 71.0401 76.8518 71.0401L77.909 71.5908C78.1205 71.5908 78.3319 71.5908 78.3319 71.5908L79.6006 71.0401C79.8121 71.0401 79.8121 71.0401 79.6006 70.8566ZM82.3495 69.5716L81.2922 69.0209C81.0808 69.0209 80.8693 69.0209 80.8693 69.0209L78.9663 69.9387C78.7548 69.9387 78.7548 70.1223 78.9663 70.1223L80.0235 70.673C80.235 70.673 80.4464 70.673 80.4464 70.673L82.3495 69.7552V69.5716ZM91.2304 70.8566L90.1732 70.3059C89.9617 70.3059 89.7503 70.3059 89.7503 70.3059L88.4816 70.8566C88.2701 70.8566 88.2701 71.0401 88.4816 71.0401L89.5388 71.5908C89.7503 71.5908 89.9617 71.5908 89.9617 71.5908L91.2304 71.0401C91.2304 70.8566 91.2304 70.8566 91.2304 70.8566ZM93.3449 69.7552L92.2877 69.2045C92.0762 69.2045 91.8648 69.2045 91.8648 69.2045L90.3846 69.7552C90.1732 69.7552 90.1732 69.9387 90.3846 69.9387L91.4419 70.4894C91.6533 70.4894 91.8648 70.4894 91.8648 70.4894L93.1335 69.9387C93.3449 69.9387 93.3449 69.7552 93.3449 69.7552ZM95.248 68.6538L94.1907 68.1031C93.9793 68.1031 93.7678 68.1031 93.7678 68.1031L92.4991 68.6538C92.2877 68.6538 92.2877 68.8374 92.4991 68.8374L93.5564 69.3881C93.7678 69.3881 93.9793 69.3881 93.9793 69.3881L95.248 68.8374C95.4595 68.8374 95.4595 68.8374 95.248 68.6538ZM93.1335 71.7744L92.0762 71.2237C91.8648 71.2237 91.6533 71.2237 91.6533 71.2237L90.3846 71.7744C90.1732 71.7744 90.1732 71.958 90.3846 71.958L91.4419 72.5087C91.6533 72.5087 91.8648 72.5087 91.8648 72.5087L93.1335 71.958V71.7744ZM95.248 70.673L93.9793 70.1223C93.7678 70.1223 93.5564 70.1223 93.5564 70.1223L92.2877 70.673C92.0762 70.673 92.0762 70.8566 92.2877 70.8566L93.3449 71.4073C93.5564 71.4073 93.7678 71.4073 93.7678 71.4073L95.0366 70.8566C95.248 70.8566 95.248 70.673 95.248 70.673ZM97.1511 69.5716L96.0938 69.0209C95.8824 69.0209 95.6709 69.0209 95.6709 69.0209L94.4022 69.7552C94.1908 69.7552 94.1908 69.9387 94.4022 69.9387L95.4595 70.4894C95.6709 70.4894 95.8824 70.4894 95.8824 70.4894L97.1511 69.9387C97.3625 69.7552 97.3625 69.7552 97.1511 69.5716ZM65.6449 79.8513L64.5876 79.3006C64.3762 79.3006 64.1647 79.3006 64.1647 79.3006L62.2616 80.2184C62.0502 80.2184 62.0502 80.402 62.2616 80.402L63.3189 80.9527C63.5303 80.9527 63.7418 80.9527 63.7418 80.9527L65.6449 80.0348C65.8563 80.0348 65.8563 79.8513 65.6449 79.8513ZM67.7594 78.7499L66.7021 78.1992C66.4907 78.1992 66.2792 78.1992 66.2792 78.1992L65.0105 78.9335C64.7991 78.9335 64.7991 79.117 65.0105 79.117L66.0678 79.6677C66.2792 79.6677 66.4907 79.6677 66.4907 79.6677L67.7594 79.117C67.7594 78.9335 67.7594 78.7499 67.7594 78.7499ZM69.8739 77.6485L68.6052 77.0978C68.3937 77.0978 68.1823 77.0978 68.1823 77.0978L66.9136 77.6485C66.7021 77.6485 66.7021 77.8321 66.9136 77.8321L67.9708 78.3828C68.1823 78.3828 68.3937 78.3828 68.3937 78.3828L69.6624 77.8321C69.8739 77.8321 69.8739 77.8321 69.8739 77.6485ZM71.7769 76.7307L70.7197 76.18C70.5082 76.18 70.2968 76.18 70.2968 76.18L69.0281 76.7307C68.8166 76.7307 68.8166 76.9142 69.0281 76.9142L70.0853 77.4649C70.2968 77.4649 70.5082 77.4649 70.5082 77.4649L71.7769 76.9142C71.9884 76.9142 71.9884 76.7307 71.7769 76.7307ZM73.8914 75.6293L72.8342 75.0786C72.6227 75.0786 72.4113 75.0786 72.4113 75.0786L71.1426 75.6293C70.9311 75.6293 70.9311 75.8128 71.1426 75.8128L72.1998 76.3635C72.4113 76.3635 72.6227 76.3635 72.6227 76.3635L73.8914 75.8128C74.1029 75.8128 74.1029 75.6293 73.8914 75.6293ZM76.006 74.7115L74.9487 74.1608C74.7373 74.1608 74.5258 74.1608 74.5258 74.1608L73.2571 74.7115C73.0456 74.7115 73.0456 74.895 73.2571 74.895L74.3144 75.4457C74.5258 75.4457 74.7373 75.4457 74.7373 75.4457L76.006 74.895C76.006 74.7114 76.006 74.7115 76.006 74.7115ZM78.1205 73.6101L77.0632 73.0594C76.8518 73.0594 76.6403 73.0594 76.6403 73.0594L75.3716 73.6101C75.1602 73.6101 75.1602 73.7936 75.3716 73.7936L76.4289 74.3443C76.6403 74.3443 76.8518 74.3443 76.8518 74.3443L78.1205 73.7936V73.6101ZM80.0235 72.5087L78.9663 71.958C78.7548 71.958 78.5434 71.958 78.5434 71.958L77.2747 72.5087C77.0632 72.5087 77.0632 72.6922 77.2747 72.6922L78.3319 73.2429C78.5434 73.2429 78.7548 73.2429 78.7548 73.2429L80.0235 72.6922C80.235 72.6922 80.235 72.6922 80.0235 72.5087ZM82.138 71.5908L81.0808 71.0401C80.8693 71.0401 80.6579 71.0401 80.6579 71.0401L79.3892 71.5908C79.1777 71.5908 79.1777 71.7744 79.3892 71.7744L80.4464 72.3251C80.6579 72.3251 80.8693 72.3251 80.8693 72.3251L82.138 71.7744C82.3495 71.5908 82.3495 71.5908 82.138 71.5908ZM84.2526 70.4894L83.1953 69.9387C82.9838 69.9387 82.7724 69.9387 82.7724 69.9387L81.5037 70.4894C81.2922 70.4894 81.2922 70.673 81.5037 70.673L82.5609 71.2237C82.7724 71.2237 82.9838 71.2237 82.9838 71.2237L84.2526 70.673C84.464 70.673 84.464 70.4894 84.2526 70.4894ZM66.9136 81.1362L65.8563 80.5855C65.6449 80.5855 65.4334 80.5855 65.4334 80.5855L64.1647 81.1362C63.9532 81.1362 63.9532 81.3198 64.1647 81.3198L65.222 81.8705C65.4334 81.8705 65.6449 81.8705 65.6449 81.8705L66.9136 81.3198V81.1362ZM68.8166 80.0348L67.7594 79.4841C67.5479 79.4841 67.3365 79.4841 67.3365 79.4841L66.0678 80.0348C65.8563 80.0348 65.8563 80.2184 66.0678 80.2184L67.125 80.7691C67.3365 80.7691 67.5479 80.7691 67.5479 80.7691L68.8166 80.2184C69.0281 80.2184 69.0281 80.0348 68.8166 80.0348ZM70.9311 78.9335L69.8739 78.3828C69.6624 78.3828 69.451 78.3828 69.451 78.3828L68.1823 78.9335C67.9708 78.9335 67.9708 79.117 68.1823 79.117L69.2395 79.6677C69.451 79.6677 69.6624 79.6677 69.6624 79.6677L70.9311 79.117C71.1426 79.117 71.1426 79.117 70.9311 78.9335ZM73.0456 78.0156L71.9884 77.4649C71.7769 77.4649 71.5655 77.4649 71.5655 77.4649L70.2968 78.0156C70.0853 78.0156 70.0853 78.1992 70.2968 78.1992L71.354 78.7499C71.5655 78.7499 71.7769 78.7499 71.7769 78.7499L73.0456 78.1992V78.0156ZM70.5082 82.9719L69.451 82.4212C69.2395 82.4212 69.0281 82.4212 69.0281 82.4212L67.7594 82.9719C67.5479 82.9719 67.5479 83.1555 67.7594 83.1555L68.8166 83.7062C69.0281 83.7062 69.2395 83.7062 69.2395 83.7062L70.5082 83.1555C70.7197 83.1555 70.7197 82.9719 70.5082 82.9719ZM72.6227 81.8705L71.5655 81.3198C71.354 81.3198 71.1426 81.3198 71.1426 81.3198L69.8739 81.8705C69.6624 81.8705 69.6624 82.0541 69.8739 82.0541L70.9311 82.6048C71.1426 82.6048 71.354 82.6048 71.354 82.6048L72.6227 82.0541C72.8342 82.0541 72.8342 82.0541 72.6227 81.8705ZM74.7373 80.9527L73.68 80.402C73.4685 80.402 73.2571 80.402 73.2571 80.402L71.9884 80.9527C71.7769 80.9527 71.7769 81.1362 71.9884 81.1362L73.0456 81.6869C73.2571 81.6869 73.4685 81.6869 73.4685 81.6869L74.7373 81.1362V80.9527ZM76.8518 79.8513L75.7945 79.3006C75.5831 79.3006 75.3716 79.3006 75.3716 79.3006L74.1029 79.8513C73.8914 79.8513 73.8914 80.0348 74.1029 80.0348L74.9487 80.7691C75.1602 80.7691 75.3716 80.7691 75.3716 80.7691L76.6403 80.2184C76.8518 80.0348 76.8518 79.8513 76.8518 79.8513ZM89.7503 73.2429L88.693 72.6922C88.4816 72.6922 88.2701 72.6922 88.2701 72.6922L86.3671 73.6101C86.1556 73.6101 86.1556 73.7936 86.3671 73.7936L87.4243 74.3443C87.6358 74.3443 87.8472 74.3443 87.8472 74.3443L89.7503 73.4265C89.9617 73.4265 89.9617 73.4265 89.7503 73.2429ZM75.1602 76.9142L74.1029 76.3635C73.8914 76.3635 73.68 76.3635 73.68 76.3635L72.1998 77.0978C71.9884 77.0978 71.9884 77.2814 72.1998 77.2814L73.2571 77.8321C73.4685 77.8321 73.68 77.8321 73.68 77.8321L74.9487 77.2814C75.1602 77.0978 75.1602 76.9142 75.1602 76.9142ZM77.0632 75.8128L76.006 75.2621C75.7945 75.2621 75.5831 75.2621 75.5831 75.2621L74.3144 75.8128C74.1029 75.8128 74.1029 75.9964 74.3144 75.9964L75.3716 76.5471C75.5831 76.5471 75.7945 76.5471 75.7945 76.5471L77.0632 75.9964C77.2747 75.9964 77.2747 75.9964 77.0632 75.8128ZM79.1777 74.895L78.1205 74.3443C77.909 74.3443 77.6976 74.3443 77.6976 74.3443L76.4289 74.895C76.2174 74.895 76.2174 75.0786 76.4289 75.0786L77.4861 75.6293C77.6976 75.6293 77.909 75.6293 77.909 75.6293L79.1777 75.0786C79.3892 75.0786 79.3892 74.895 79.1777 74.895ZM81.2922 73.7936L80.235 73.2429C80.0235 73.2429 79.8121 73.2429 79.8121 73.2429L78.5434 73.7936C78.3319 73.7936 78.3319 73.9772 78.5434 73.9772L79.6006 74.5279C79.8121 74.5279 80.0235 74.5279 80.0235 74.5279L81.2922 73.9772C81.2922 73.9772 81.2922 73.9772 81.2922 73.7936ZM83.4067 72.8758L82.3495 72.3251C82.138 72.3251 81.9266 72.3251 81.9266 72.3251L80.6579 72.8758C80.4464 72.8758 80.4464 73.0594 80.6579 73.0594L81.7151 73.6101C81.9266 73.6101 82.138 73.6101 82.138 73.6101L83.4067 73.0594C83.4067 72.8758 83.4067 72.8758 83.4067 72.8758ZM86.1556 71.4073L85.0984 70.8566C84.8869 70.8566 84.6754 70.8566 84.6754 70.8566L82.7724 71.7744C82.5609 71.7744 82.5609 71.958 82.7724 71.958L83.8296 72.5087C84.0411 72.5087 84.2526 72.5087 84.2526 72.5087L86.1556 71.5908V71.4073ZM69.451 81.6869L68.3937 81.1362C68.1823 81.1362 67.9708 81.1362 67.9708 81.1362L66.0678 82.0541C65.8563 82.0541 65.8563 82.2376 66.0678 82.2376L67.125 82.7883C67.3365 82.7883 67.5479 82.7883 67.5479 82.7883L69.451 81.8705V81.6869ZM71.5655 80.5855L70.5082 80.0348C70.2968 80.0348 70.0853 80.0348 70.0853 80.0348L68.6052 80.7691C68.3937 80.7691 68.3937 80.9527 68.6052 80.9527L69.6624 81.5034C69.8739 81.5034 70.0853 81.5034 70.0853 81.5034L71.354 80.9527C71.5655 80.7691 71.5655 80.7691 71.5655 80.5855ZM73.4685 79.6677L72.4113 79.117C72.1998 79.117 71.9884 79.117 71.9884 79.117L70.7197 79.6677C70.5082 79.6677 70.5082 79.8513 70.7197 79.8513L71.7769 80.402C71.9884 80.402 72.1998 80.402 72.1998 80.402L73.4685 79.8513C73.68 79.8513 73.68 79.6677 73.4685 79.6677ZM75.5831 78.5663L74.5258 78.0156C74.3143 78.0156 74.1029 78.0156 74.1029 78.0156L72.8342 78.7499C72.6227 78.7499 72.6227 78.9335 72.8342 78.9335L73.8914 79.4841C74.1029 79.4841 74.3144 79.4841 74.3144 79.4841L75.5831 78.9335C75.7945 78.7499 75.7945 78.5663 75.5831 78.5663ZM77.6976 77.4649L76.6403 76.9142C76.4289 76.9142 76.2174 76.9142 76.2174 76.9142L74.9487 77.6485C74.7373 77.6485 74.7373 77.8321 74.9487 77.8321L76.006 78.3828C76.2174 78.3828 76.4289 78.3828 76.4289 78.3828L77.6976 77.8321C77.6976 77.6485 77.6976 77.6485 77.6976 77.4649ZM87.2129 74.7115L86.1556 74.1608C85.9442 74.1608 85.7327 74.1608 85.7327 74.1608L76.006 78.9335C75.7945 78.9335 75.7945 79.117 76.006 79.117L77.0632 79.6677C77.2747 79.6677 77.4861 79.6677 77.4861 79.6677L87.0014 74.895C87.2129 74.895 87.2129 74.7115 87.2129 74.7115ZM79.8121 76.5471L78.7548 75.9964C78.5434 75.9964 78.3319 75.9964 78.3319 75.9964L77.0632 76.5471C76.8518 76.5471 76.8518 76.7307 77.0632 76.7307L78.1205 77.2814C78.3319 77.2814 78.5434 77.2814 78.5434 77.2814L79.8121 76.7307V76.5471ZM81.7151 75.4457L80.6579 74.895C80.4464 74.895 80.235 74.895 80.235 74.895L78.9663 75.4457C78.7548 75.4457 78.7548 75.6293 78.9663 75.6293L80.0235 76.18C80.235 76.18 80.4464 76.18 80.4464 76.18L81.7151 75.6293C81.9266 75.6293 81.9266 75.4457 81.7151 75.4457ZM83.8296 74.3443L82.7724 73.7936C82.5609 73.7936 82.3495 73.7936 82.3495 73.7936L81.0808 74.3443C80.8693 74.3443 80.8693 74.5279 81.0808 74.5279L82.138 75.0786C82.3495 75.0786 82.5609 75.0786 82.5609 75.0786L83.8296 74.5279C84.0411 74.5279 84.0411 74.5279 83.8296 74.3443ZM85.9442 73.4265L84.8869 72.8758C84.6755 72.8758 84.464 72.8758 84.464 72.8758L83.1953 73.4265C82.9838 73.4265 82.9838 73.6101 83.1953 73.6101L84.2526 74.1608C84.464 74.1608 84.6754 74.1608 84.6754 74.1608L85.9442 73.6101C86.1556 73.6101 86.1556 73.4265 85.9442 73.4265ZM88.0587 72.3251L87.0014 71.7744C86.79 71.7744 86.5785 71.7744 86.5785 71.7744L85.3098 72.3251C85.0984 72.3251 85.0984 72.5087 85.3098 72.5087L86.3671 73.0594C86.5785 73.0594 86.79 73.0594 86.79 73.0594L88.0587 72.5087C88.0587 72.5087 88.0587 72.5087 88.0587 72.3251ZM85.5213 67.9195L84.464 67.3688C84.2525 67.3688 84.0411 67.3688 84.0411 67.3688L82.7724 67.9195C82.5609 67.9195 82.5609 68.1031 82.7724 68.1031L83.8296 68.6538C84.0411 68.6538 84.2526 68.6538 84.2526 68.6538L85.5213 68.1031C85.7327 68.1031 85.7327 67.9195 85.5213 67.9195ZM87.6358 66.8181L86.5785 66.2674C86.3671 66.2674 86.1556 66.2674 86.1556 66.2674L84.8869 66.8181C84.6755 66.8181 84.6755 67.0017 84.8869 67.0017L85.9442 67.5524C86.1556 67.5524 86.3671 67.5524 86.3671 67.5524L87.6358 67.0017C87.6358 67.0017 87.6358 67.0017 87.6358 66.8181ZM89.7503 65.9003L88.693 65.3496C88.4816 65.3496 88.2701 65.3496 88.2701 65.3496L87.0014 65.9003C86.79 65.9003 86.79 66.0839 87.0014 66.0839L88.0587 66.6346C88.2701 66.6346 88.4816 66.6346 88.4816 66.6346L89.7503 66.0839V65.9003ZM91.6533 66.8181L90.5961 66.2674C90.3846 66.2674 90.1732 66.2674 90.1732 66.2674L88.9045 66.8181C88.693 66.8181 88.693 67.0017 88.9045 67.0017L89.9617 67.5524C90.1732 67.5524 90.3846 67.5524 90.3846 67.5524L91.6533 67.0017V66.8181ZM93.3449 67.736L92.2877 67.1853C92.2877 67.1853 92.0762 67.1853 91.8648 67.1853L89.9617 68.1031H89.7503L88.2701 67.3688C88.0587 67.3688 87.8472 67.3688 87.6358 67.3688L84.464 69.0209C84.2525 69.0209 84.2525 69.2045 84.464 69.3881L87.2129 70.8566C87.4243 70.8566 87.6358 70.8566 87.8472 70.8566L93.1335 68.1031C93.5564 67.9195 93.5564 67.736 93.3449 67.736Z"
-          fill="#263144"
-        />
-        <path
-          d="M63.1074 79.3008L61.8387 79.8515C61.6273 79.8515 61.4158 79.8515 61.4158 79.8515L60.1471 79.3008C59.9357 79.3008 59.9357 79.1172 60.1471 79.1172L61.4158 78.5665C61.6273 78.5665 61.8387 78.5665 61.8387 78.5665L62.896 79.1172C63.1074 79.1172 63.1074 79.1172 63.1074 79.3008ZM65.222 78.0158L64.1647 77.4651C63.9532 77.4651 63.7418 77.4651 63.7418 77.4651L62.2616 78.0158C62.0502 78.0158 62.0502 78.1994 62.2616 78.1994L63.3189 78.7501C63.5303 78.7501 63.7418 78.7501 63.7418 78.7501L65.0105 78.1994C65.222 78.1994 65.222 78.0158 65.222 78.0158ZM67.125 76.9144L66.0678 76.3637C65.8563 76.3637 65.6449 76.3637 65.6449 76.3637L64.3762 77.098C64.1647 77.098 64.1647 77.2816 64.3762 77.2816L65.4334 77.8323C65.6449 77.8323 65.8563 77.8323 65.8563 77.8323L67.125 77.2816C67.3365 77.098 67.3365 77.098 67.125 76.9144ZM69.2395 75.9966L68.1823 75.4459C67.9708 75.4459 67.7594 75.4459 67.7594 75.4459L66.4907 75.9966C66.2792 75.9966 66.2792 76.1802 66.4907 76.1802L67.5479 76.7309C67.7594 76.7309 67.9708 76.7309 67.9708 76.7309L69.2395 76.1802C69.451 76.1802 69.451 75.9966 69.2395 75.9966ZM71.354 74.8952L70.2968 74.3445C70.0853 74.3445 69.8739 74.3445 69.8739 74.3445L68.6052 75.0788C68.3937 75.0788 68.3937 75.2624 68.6052 75.2624L69.6624 75.8131C69.8739 75.8131 70.0853 75.8131 70.0853 75.8131L71.354 75.2624C71.354 75.0788 71.354 74.8952 71.354 74.8952ZM73.4685 73.7938L72.4113 73.2431C72.1998 73.2431 71.9884 73.2431 71.9884 73.2431L70.7197 73.7938C70.5082 73.7938 70.5082 73.9774 70.7197 73.9774L71.7769 74.5281C71.9884 74.5281 72.1998 74.5281 72.1998 74.5281L73.4685 73.9774C73.4685 73.9774 73.4685 73.9774 73.4685 73.7938ZM75.3716 72.876L74.3144 72.3253C74.1029 72.3253 73.8914 72.3253 73.8914 72.3253L72.6227 72.876C72.4113 72.876 72.4113 73.0596 72.6227 73.0596L73.68 73.6103C73.8914 73.6103 74.1029 73.6103 74.1029 73.6103L75.3716 73.0596C75.5831 73.0596 75.5831 72.876 75.3716 72.876ZM77.4861 71.7746L76.4289 71.2239C76.2174 71.2239 76.006 71.2239 76.006 71.2239L74.7373 71.7746C74.5258 71.7746 74.5258 71.9582 74.7373 71.9582L75.7945 72.5089C76.006 72.5089 76.2174 72.5089 76.2174 72.5089L77.4861 71.9582C77.6976 71.9582 77.6976 71.7746 77.4861 71.7746ZM79.6006 70.6732L78.5434 70.1225C78.3319 70.1225 78.1205 70.1225 78.1205 70.1225L76.8518 70.6732C76.6403 70.6732 76.6403 70.8568 76.8518 70.8568L77.909 71.4075C78.1205 71.4075 78.3319 71.4075 78.3319 71.4075L79.6006 70.8568C79.8121 70.8568 79.8121 70.8568 79.6006 70.6732ZM82.3495 69.3883L81.2922 68.8376C81.0808 68.8376 80.8693 68.8376 80.8693 68.8376L78.7548 69.7554C78.5434 69.7554 78.5434 69.939 78.7548 69.939L79.8121 70.4897C80.0235 70.4897 80.235 70.4897 80.235 70.4897L82.138 69.5718C82.3495 69.5718 82.3495 69.3883 82.3495 69.3883ZM91.2304 70.6732L90.1732 70.1225C89.9617 70.1225 89.7503 70.1225 89.7503 70.1225L88.4816 70.6732C88.2701 70.6732 88.2701 70.8568 88.4816 70.8568L89.5388 71.4075C89.7503 71.4075 89.9617 71.4075 89.9617 71.4075L91.2304 70.8568V70.6732ZM93.3449 69.5718L92.2877 69.0211C92.0762 69.0211 91.8648 69.0211 91.8648 69.0211L90.3846 69.7554C90.1732 69.7554 90.1732 69.939 90.3846 69.939L91.4419 70.4897C91.6533 70.4897 91.8648 70.4897 91.8648 70.4897L93.1335 69.939C93.3449 69.7554 93.3449 69.5718 93.3449 69.5718ZM95.248 68.4704L94.1907 67.9197C93.9793 67.9197 93.7678 67.9197 93.7678 67.9197L92.4991 68.4704C92.2877 68.4704 92.2877 68.654 92.4991 68.654L93.5564 69.2047C93.7678 69.2047 93.9793 69.2047 93.9793 69.2047L95.248 68.654C95.4595 68.654 95.4595 68.654 95.248 68.4704ZM93.1335 71.591L92.0762 71.0404C91.8648 71.0404 91.6533 71.0404 91.6533 71.0404L90.1732 71.591C89.9617 71.591 89.9617 71.7746 90.1732 71.7746L91.2304 72.3253C91.4419 72.3253 91.6533 72.3253 91.6533 72.3253L92.922 71.7746C93.1335 71.7746 93.1335 71.591 93.1335 71.591ZM95.248 70.4897L93.9793 69.939C93.7678 69.939 93.5564 69.939 93.5564 69.939L92.2877 70.4897C92.0762 70.4897 92.0762 70.6732 92.2877 70.6732L93.3449 71.2239C93.5564 71.2239 93.7678 71.2239 93.7678 71.2239L95.0366 70.6732C95.248 70.6732 95.248 70.6732 95.248 70.4897ZM97.1511 69.5718L96.0938 69.0211C95.8824 69.0211 95.6709 69.0211 95.6709 69.0211L94.4022 69.5718C94.1908 69.5718 94.1908 69.7554 94.4022 69.7554L95.4595 70.3061C95.6709 70.3061 95.8824 70.3061 95.8824 70.3061L97.1511 69.7554C97.3625 69.5718 97.3625 69.5718 97.1511 69.5718ZM65.6449 79.6679L64.5876 79.1172C64.3762 79.1172 64.1647 79.1172 64.1647 79.1172L62.2616 80.0351C62.0502 80.0351 62.0502 80.2186 62.2616 80.2186L63.3189 80.7693C63.5303 80.7693 63.7418 80.7693 63.7418 80.7693L65.6449 79.8515C65.8563 79.8515 65.8563 79.6679 65.6449 79.6679ZM67.7594 78.5665L66.7021 78.0158C66.4907 78.0158 66.2792 78.0158 66.2792 78.0158L65.0105 78.5665C64.7991 78.5665 64.7991 78.7501 65.0105 78.7501L66.0678 79.3008C66.2792 79.3008 66.4907 79.3008 66.4907 79.3008L67.7594 78.7501C67.7594 78.7501 67.7594 78.7501 67.7594 78.5665ZM69.8739 77.6487L68.6052 77.098C68.3937 77.098 68.1823 77.098 68.1823 77.098L66.9136 77.6487C66.7021 77.6487 66.7021 77.8323 66.9136 77.8323L67.9708 78.383C68.1823 78.383 68.3937 78.383 68.3937 78.383L69.6624 77.8323C69.8739 77.6487 69.8739 77.6487 69.8739 77.6487ZM71.7769 76.5473L70.7197 75.9966C70.5082 75.9966 70.2968 75.9966 70.2968 75.9966L69.0281 76.5473C68.8166 76.5473 68.8166 76.7309 69.0281 76.7309L70.0853 77.2816C70.2968 77.2816 70.5082 77.2816 70.5082 77.2816L71.7769 76.7309C71.9884 76.7309 71.9884 76.5473 71.7769 76.5473ZM73.8914 75.4459L72.8342 74.8952C72.6227 74.8952 72.4113 74.8952 72.4113 74.8952L71.1426 75.4459C70.9311 75.4459 70.9311 75.6295 71.1426 75.6295L72.1998 76.1802C72.4113 76.1802 72.6227 76.1802 72.6227 76.1802L73.8914 75.6295C74.1029 75.6295 74.1029 75.6295 73.8914 75.4459ZM76.006 74.5281L74.9487 73.9774C74.7373 73.9774 74.5258 73.9774 74.5258 73.9774L73.2571 74.5281C73.0456 74.5281 73.0456 74.7117 73.2571 74.7117L74.3144 75.2624C74.5258 75.2624 74.7373 75.2624 74.7373 75.2624L76.006 74.7117V74.5281ZM78.1205 73.4267L77.0632 72.876C76.8518 72.876 76.6403 72.876 76.6403 72.876L75.3716 73.4267C75.1602 73.4267 75.1602 73.6103 75.3716 73.6103L76.4289 74.161C76.6403 74.161 76.8518 74.161 76.8518 74.161L78.1205 73.6103V73.4267ZM80.0235 72.3253L78.9663 71.7746C78.7548 71.7746 78.5434 71.7746 78.5434 71.7746L77.2747 72.3253C77.0632 72.3253 77.0632 72.5089 77.2747 72.5089L78.3319 73.0596C78.5434 73.0596 78.7548 73.0596 78.7548 73.0596L80.0235 72.5089C80.235 72.5089 80.235 72.5089 80.0235 72.3253ZM82.138 71.4075L81.0808 70.8568C80.8693 70.8568 80.6579 70.8568 80.6579 70.8568L79.3892 71.4075C79.1777 71.5911 79.1777 71.591 79.3892 71.591L80.4464 72.1417C80.6579 72.1417 80.8693 72.1417 80.8693 72.1417L82.138 71.591C82.3495 71.591 82.3495 71.4075 82.138 71.4075ZM84.2526 70.3061L83.1953 69.7554C82.9838 69.7554 82.7724 69.7554 82.7724 69.7554L81.5037 70.3061C81.2922 70.3061 81.2922 70.4897 81.5037 70.4897L82.5609 71.0404C82.7724 71.0404 82.9838 71.0404 82.9838 71.0404L84.2526 70.4897C84.464 70.4897 84.464 70.3061 84.2526 70.3061ZM66.9136 80.9529L65.8563 80.4022C65.6449 80.4022 65.4334 80.4022 65.4334 80.4022L64.1647 80.9529C63.9532 80.9529 63.9532 81.1365 64.1647 81.1365L65.222 81.6871C65.4334 81.6871 65.6449 81.6871 65.6449 81.6871L66.9136 81.1365V80.9529ZM68.8166 79.8515L67.7594 79.3008C67.5479 79.3008 67.3365 79.3008 67.3365 79.3008L66.0678 79.8515C65.8563 79.8515 65.8563 80.0351 66.0678 80.0351L67.125 80.5858C67.3365 80.5858 67.5479 80.5858 67.5479 80.5858L68.8166 80.0351C69.0281 80.0351 69.0281 79.8515 68.8166 79.8515ZM70.9311 78.9337L69.8739 78.383C69.6624 78.383 69.451 78.383 69.451 78.383L68.1823 78.9337C67.9708 78.9337 67.9708 79.1172 68.1823 79.1172L69.2395 79.6679C69.451 79.6679 69.6624 79.6679 69.6624 79.6679L70.9311 79.1172C71.1426 78.9337 71.1426 78.9337 70.9311 78.9337ZM73.0456 77.8323L71.9884 77.2816C71.7769 77.2816 71.5655 77.2816 71.5655 77.2816L70.2968 77.8323C70.0853 77.8323 70.0853 78.0158 70.2968 78.0158L71.354 78.5665C71.5655 78.5665 71.7769 78.5665 71.7769 78.5665L73.0456 78.0158V77.8323ZM70.5082 82.7885L69.451 82.2378C69.2395 82.2378 69.0281 82.2378 69.0281 82.2378L67.7594 82.7885C67.5479 82.7885 67.5479 82.9721 67.7594 82.9721L68.8166 83.5228C69.0281 83.5228 69.2395 83.5228 69.2395 83.5228L70.5082 82.9721C70.7197 82.9721 70.7197 82.7885 70.5082 82.7885ZM72.6227 81.6871L71.5655 81.1365C71.354 81.1365 71.1426 81.1365 71.1426 81.1365L69.8739 81.6871C69.6624 81.6871 69.6624 81.8707 69.8739 81.8707L70.9311 82.4214C71.1426 82.4214 71.354 82.4214 71.354 82.4214L72.6227 81.8707C72.8342 81.8707 72.8342 81.8707 72.6227 81.6871ZM74.7373 80.7693L73.68 80.2186C73.4685 80.2186 73.2571 80.2186 73.2571 80.2186L71.9884 80.7693C71.7769 80.7693 71.7769 80.9529 71.9884 80.9529L73.0456 81.5036C73.2571 81.5036 73.4685 81.5036 73.4685 81.5036L74.7373 80.9529V80.7693ZM76.8518 79.6679L75.7945 79.1172C75.5831 79.1172 75.3716 79.1172 75.3716 79.1172L74.1029 79.6679C73.8914 79.6679 73.8914 79.8515 74.1029 79.8515L75.1602 80.4022C75.3716 80.4022 75.5831 80.4022 75.5831 80.4022L76.8518 79.8515V79.6679ZM89.7503 73.0596L88.693 72.5089C88.4816 72.5089 88.2701 72.5089 88.2701 72.5089L86.3671 73.4267C86.1556 73.4267 86.1556 73.6103 86.3671 73.6103L87.4243 74.161C87.6358 74.161 87.8472 74.161 87.8472 74.161L89.7503 73.4267C89.9617 73.2431 89.9617 73.2431 89.7503 73.0596ZM75.1602 76.7309L74.1029 76.1802C73.8914 76.1802 73.68 76.1802 73.68 76.1802L72.4113 76.7309C72.1998 76.7309 72.1998 76.9144 72.4113 76.9144L73.4685 77.4651C73.68 77.4651 73.8914 77.4651 73.8914 77.4651L75.1602 76.9144C75.1602 76.9144 75.1602 76.9144 75.1602 76.7309ZM77.0632 75.8131L76.006 75.2624C75.7945 75.2624 75.5831 75.2624 75.5831 75.2624L74.3144 75.8131C74.1029 75.8131 74.1029 75.9966 74.3144 75.9966L75.3716 76.5473C75.5831 76.5473 75.7945 76.5473 75.7945 76.5473L77.0632 75.9966C77.2747 75.8131 77.2747 75.8131 77.0632 75.8131ZM79.1777 74.7117L78.1205 74.161C77.909 74.161 77.6976 74.161 77.6976 74.161L76.4289 74.7117C76.2174 74.7117 76.2174 74.8952 76.4289 74.8952L77.4861 75.4459C77.6976 75.4459 77.909 75.4459 77.909 75.4459L79.1777 74.8952C79.3892 74.8952 79.3892 74.7117 79.1777 74.7117ZM81.2922 73.6103L80.235 73.0596C80.0235 73.0596 79.8121 73.0596 79.8121 73.0596L78.5434 73.6103C78.3319 73.6103 78.3319 73.7938 78.5434 73.7938L79.6006 74.3445C79.8121 74.3445 80.0235 74.3445 80.0235 74.3445L81.2922 73.7938C81.2922 73.7938 81.2922 73.7938 81.2922 73.6103ZM83.4067 72.6924L82.3495 72.1417C82.138 72.1417 81.9266 72.1417 81.9266 72.1417L80.6579 72.6924C80.4464 72.6924 80.4464 72.876 80.6579 72.876L81.7151 73.4267C81.9266 73.4267 82.138 73.4267 82.138 73.4267L83.4067 72.876V72.6924ZM86.1556 71.2239L85.0984 70.6732C84.8869 70.6732 84.6754 70.6732 84.6754 70.6732L82.7724 71.591C82.5609 71.591 82.5609 71.7746 82.7724 71.7746L83.8296 72.3253C84.0411 72.3253 84.2526 72.3253 84.2526 72.3253L86.1556 71.4075C86.1556 71.4075 86.1556 71.4075 86.1556 71.2239ZM69.451 81.5036L68.3937 80.9529C68.1823 80.9529 67.9708 80.9529 67.9708 80.9529L66.0678 81.8707C65.8563 81.8707 65.8563 82.0543 66.0678 82.0543L67.125 82.605C67.3365 82.605 67.5479 82.605 67.5479 82.605L69.451 81.6871V81.5036ZM71.5655 80.4022L70.5082 79.8515C70.2968 79.8515 70.0853 79.8515 70.0853 79.8515L68.6052 80.5858V80.7693L69.6624 81.32C69.8739 81.32 70.0853 81.32 70.0853 81.32L71.354 80.7693C71.5655 80.5858 71.5655 80.5858 71.5655 80.4022ZM73.4685 79.4844L72.4113 78.9337C72.1998 78.9337 71.9884 78.9337 71.9884 78.9337L70.7197 79.4844C70.5082 79.4844 70.5082 79.6679 70.7197 79.6679L71.7769 80.2186C71.9884 80.2186 72.1998 80.2186 72.1998 80.2186L73.4685 79.6679C73.68 79.6679 73.68 79.4844 73.4685 79.4844ZM75.5831 78.383L74.5258 77.8323C74.3143 77.8323 74.1029 77.8323 74.1029 77.8323L72.8342 78.5665C72.6227 78.5665 72.6227 78.7501 72.8342 78.7501L73.8914 79.3008C74.1029 79.3008 74.3144 79.3008 74.3144 79.3008L75.5831 78.7501C75.7945 78.5665 75.7945 78.383 75.5831 78.383ZM77.6976 77.4651L76.6403 76.9144C76.4289 76.9144 76.2174 76.9144 76.2174 76.9144L74.9487 77.4651C74.7373 77.4651 74.7373 77.6487 74.9487 77.6487L76.006 78.1994C76.2174 78.1994 76.4289 78.1994 76.4289 78.1994L77.6976 77.6487C77.6976 77.4651 77.6976 77.4651 77.6976 77.4651ZM87.2129 74.5281L86.1556 73.9774C85.9442 73.9774 85.7327 73.9774 85.7327 73.9774L76.2174 78.7501C76.006 78.7501 76.006 78.9337 76.2174 78.9337L77.2747 79.4844C77.4861 79.4844 77.6976 79.4844 77.6976 79.4844L87.2129 74.7117V74.5281ZM79.8121 76.3637L78.7548 75.8131C78.5434 75.8131 78.3319 75.8131 78.3319 75.8131L77.0632 76.3637C76.8518 76.3637 76.8518 76.5473 77.0632 76.5473L78.1205 77.098C78.3319 77.098 78.5434 77.098 78.5434 77.098L79.8121 76.5473V76.3637ZM81.7151 75.2624L80.6579 74.7117C80.4464 74.7117 80.235 74.7117 80.235 74.7117L78.9663 75.2624C78.7548 75.2624 78.7548 75.4459 78.9663 75.4459L80.0235 75.9966C80.235 75.9966 80.4464 75.9966 80.4464 75.9966L81.7151 75.4459C81.9266 75.4459 81.9266 75.4459 81.7151 75.2624ZM83.8296 74.3445L82.7724 73.7938C82.5609 73.7938 82.3495 73.7938 82.3495 73.7938L81.0808 74.3445C80.8693 74.3445 80.8693 74.5281 81.0808 74.5281L82.138 75.0788C82.3495 75.0788 82.5609 75.0788 82.5609 75.0788L83.8296 74.5281C84.0411 74.3445 84.0411 74.3445 83.8296 74.3445ZM85.9442 73.2431L84.8869 72.6924C84.6755 72.6924 84.464 72.6924 84.464 72.6924L83.1953 73.2431C82.9838 73.2431 82.9838 73.4267 83.1953 73.4267L84.2526 73.9774C84.464 73.9774 84.6754 73.9774 84.6754 73.9774L85.9442 73.4267C86.1556 73.4267 86.1556 73.2431 85.9442 73.2431ZM88.0587 72.1417L87.0014 71.591C86.79 71.591 86.5785 71.591 86.5785 71.591L85.3098 72.1417C85.0984 72.1417 85.0984 72.3253 85.3098 72.3253L86.3671 72.876C86.5785 72.876 86.79 72.876 86.79 72.876L88.0587 72.3253C88.0587 72.3253 88.0587 72.3253 88.0587 72.1417ZM85.5213 67.7362L84.464 67.1855C84.2525 67.1855 84.0411 67.1855 84.0411 67.1855L82.7724 67.9197C82.5609 67.9197 82.5609 68.1033 82.7724 68.1033L83.8296 68.654C84.0411 68.654 84.2526 68.654 84.2526 68.654L85.5213 67.9197C85.7327 67.9197 85.7327 67.9197 85.5213 67.7362ZM87.6358 66.8184L86.5785 66.2676C86.3671 66.2676 86.1556 66.2676 86.1556 66.2676L84.8869 66.8184C84.6755 66.8184 84.6755 67.0019 84.8869 67.0019L85.9442 67.5526C86.1556 67.5526 86.3671 67.5526 86.3671 67.5526L87.6358 67.0019C87.6358 66.8183 87.6358 66.8184 87.6358 66.8184ZM89.7503 65.717L88.693 65.1663C88.4816 65.1663 88.2701 65.1663 88.2701 65.1663L87.0014 65.717C86.79 65.717 86.79 65.9005 87.0014 65.9005L88.0587 66.4512C88.2701 66.4512 88.4816 66.4512 88.4816 66.4512L89.7503 65.9005V65.717ZM91.6533 66.6348L90.3846 66.0841C90.1732 66.0841 89.9617 66.0841 89.9617 66.0841L88.693 66.6348C88.4816 66.6348 88.4816 66.8184 88.693 66.8184L89.7503 67.369C89.9617 67.369 90.1732 67.369 90.1732 67.369L91.4419 66.8184C91.6533 66.8184 91.6533 66.6348 91.6533 66.6348ZM93.3449 67.5526L92.2877 67.0019C92.2877 67.0019 92.0762 67.0019 91.8648 67.0019L90.1732 67.9197H89.9617L88.4816 67.1855C88.2701 67.1855 88.0587 67.1855 87.8472 67.1855L84.6754 68.8376C84.464 68.8376 84.464 69.0211 84.6754 69.2047L87.4243 70.6732C87.6358 70.6732 87.8472 70.6732 88.0587 70.6732L93.3449 67.9197C93.5564 67.7362 93.5564 67.5526 93.3449 67.5526Z"
-          fill="#364157"
-        />
-        <path
-          d="M120.411 142.263L83.8297 160.987L105.821 172.001L142.613 153.46L120.411 142.263Z"
-          fill="#F9F9FB"
-        />
-        <path
-          d="M143.036 153.828C143.036 153.644 142.824 153.644 142.824 153.644C142.824 153.644 142.824 153.644 142.613 153.461L105.82 172.001L106.032 175.489L142.824 156.948C143.458 156.581 143.458 154.562 143.036 153.828Z"
-          fill="#DDDDDF"
-        />
-        <path
-          d="M119.565 141.529C119.142 141.161 118.719 140.611 118.507 139.877C118.084 139.142 117.873 138.224 117.873 137.49V110.69C117.873 109.588 118.296 109.038 118.93 108.67L81.0804 125.558C80.4461 125.926 79.8117 128.128 79.8117 129.23L80.119 157.052C80.119 157.786 80.3305 158.704 80.7534 159.438C80.9648 159.989 81.3877 160.539 81.8106 161.09C82.2335 161.457 82.445 161.641 82.8679 161.824L120.833 142.263C120.199 142.079 119.776 141.896 119.565 141.529Z"
-          fill="#DDDDE1"
-        />
-        <path
-          d="M120.622 105.55C119.142 104.815 117.873 104.815 116.816 105.366L80.0231 124.744C81.0804 124.193 82.3491 124.193 83.8292 124.928L103.494 134.84L140.287 115.279L120.622 105.55Z"
-          fill="white"
-        />
-        <path
-          d="M126.966 58.9245C125.274 60.0259 127.389 69.3878 119.354 71.7741C118.508 71.9577 117.028 71.407 116.393 71.2234C114.913 70.8563 101.803 66.6343 101.803 66.6343C101.169 66.2671 99.4771 64.6151 97.997 63.8808C96.9397 63.3301 95.8825 62.963 94.8252 62.7794C94.6138 62.7794 93.9794 62.7794 93.5565 62.7794C90.5962 63.8808 88.0588 66.0836 87.6359 66.4507C87.4244 66.6343 87.213 66.8178 87.4244 67.0014C87.4244 67.185 87.6359 67.185 87.8473 67.185C87.6359 67.7357 88.2702 67.9192 88.4817 67.9192C88.9046 67.9192 89.7504 67.5521 89.7504 67.5521C89.9618 67.5521 90.1733 67.5521 90.3847 67.5521C92.2878 67.9192 94.4023 68.2864 94.8252 68.6535C95.4596 69.0206 93.9794 69.7549 93.9794 69.7549C93.9794 69.7549 92.0763 69.5713 91.8649 69.5713C91.2305 69.5713 90.3847 69.9385 90.5962 70.3056C90.5962 70.6727 91.0191 70.6727 91.2305 70.8563C91.8649 71.0398 93.768 71.407 93.768 71.407C94.4023 71.2234 98.6313 70.6727 98.6313 70.6727C98.6313 70.6727 130.349 78.9332 130.772 78.9332C131.829 79.1167 135.212 74.7112 135.635 70.8563C136.27 67.0014 132.675 55.4368 126.966 58.9245Z"
-          fill="#FFCCAA"
-        />
-        <path
-          d="M130.983 72.8753C132.886 72.6917 135.212 72.141 136.481 71.4068C136.481 68.2861 136.481 61.1271 135.846 58.19C134.155 51.5817 125.274 49.1953 123.582 58.5572C122.948 61.4942 122.525 71.0396 122.525 71.0396C124.851 72.6917 128.023 73.0589 130.983 72.8753Z"
-          fill="#FB6D3A"
-        />
-        <path
-          d="M130.56 104.449C130.56 104.449 98.4196 98.9416 95.4593 98.0238C86.3669 95.6375 81.715 104.265 82.1379 105.734C76.2172 104.816 68.3935 106.101 65.8561 117.298L69.0279 115.646C73.2569 118.032 92.7104 124.641 99.2654 127.394C107.089 130.515 129.503 122.622 130.349 114.912C130.349 114.912 130.56 105.55 130.56 104.449Z"
-          fill="#1E2940"
-        />
-        <path
-          d="M130.137 52.8666C122.525 52.683 106.666 61.4942 102.86 64.0641C102.86 64.0641 103.072 73.6095 102.014 83.7056C101.591 89.029 98.6311 106.651 97.9968 110.506C97.9968 111.057 97.9968 111.791 98.6311 112.158C106.032 118.216 126.12 115.462 130.772 105.733C130.772 95.8209 135.635 53.4173 130.137 52.8666Z"
-          fill="#FB6D3A"
-        />
-        <path
-          d="M124.851 54.7024C124.639 54.5188 124.639 54.3353 124.639 53.9681C124.639 53.4174 124.639 52.8667 124.639 52.316C124.639 51.5818 124.639 50.664 124.428 49.9297C124.217 48.8283 122.736 49.1954 121.679 49.0119C119.988 48.4612 118.296 48.4612 116.393 48.2776C116.181 48.2776 115.759 48.2776 115.547 48.094C115.124 47.9105 114.913 47.7269 114.49 47.7269C114.278 47.7269 113.855 47.7269 113.433 47.5433C113.221 47.5433 111.107 46.9926 111.107 46.8091C111.741 49.379 112.375 55.4367 112.375 58.5573C112.375 58.5573 109.204 60.5765 108.781 60.7601C111.952 63.1464 119.565 61.1272 123.794 58.1901C124.428 57.6394 125.274 57.0887 125.485 56.3545C125.697 55.6202 125.062 55.2531 124.851 54.7024Z"
-          fill="#FFCCAA"
-        />
-        <path
-          d="M109.204 69.9385C107.512 66.2671 102.86 65.3493 99.8998 69.2042C98.6311 70.8563 97.5738 75.9961 94.4021 83.8894L93.1334 83.3387C81.5036 77.8318 80.8692 78.0153 80.4463 76.7304C80.0234 75.8125 79.3891 73.9769 78.9662 73.6098C78.9662 73.6098 77.2745 72.1412 76.8516 71.9577C76.6402 71.7741 76.2173 71.7741 75.7944 71.9577C75.5829 72.1412 75.5829 72.3248 75.5829 72.5084C75.3715 72.5084 74.9486 72.3248 74.7371 72.3248C71.3539 72.5084 69.4509 74.7112 69.2394 74.8947C68.8165 75.0783 68.8165 75.0783 68.8165 75.4454C68.8165 75.629 69.028 75.629 69.2394 75.8125C68.8165 76.1797 69.2394 76.5468 69.6623 76.5468C70.0852 76.5468 70.931 76.5468 70.931 76.5468C71.1425 76.5468 71.3539 76.5468 71.5654 76.7304C73.257 77.6482 73.8913 77.8318 74.1028 78.3825C79.6005 83.7059 95.8822 94.1691 99.2655 92.3334C104.552 89.3964 111.318 74.7112 109.204 69.9385Z"
-          fill="#FFCCAA"
-        />
-        <path
-          d="M101.803 84.4399C103.706 84.6235 106.032 84.6235 107.512 84.0728C108.358 81.1357 110.261 73.9767 110.261 70.6725C110.473 63.697 102.437 59.6586 98.4198 68.6533C97.1511 71.4067 94.1908 80.7686 94.1908 80.7686C95.8824 82.9714 98.8427 84.0728 101.803 84.4399Z"
-          fill="#FB6D3A"
-        />
-        <path
-          d="M106.455 98.024L105.82 110.69L106.243 137.123C106.243 137.674 106.032 138.041 105.82 138.225L142.613 119.501C142.824 119.318 143.036 118.95 143.036 118.4V79.4839L106.455 98.024Z"
-          fill="#DDDDDF"
-        />
-        <path
-          d="M143.247 79.3001C143.247 79.1165 143.036 79.1165 143.036 78.9329C142.824 78.5658 142.401 78.3822 141.978 78.1987C141.344 78.0151 140.921 77.8315 140.71 77.8315L103.706 96.3717L106.455 98.0237L143.247 79.4836V79.3001Z"
-          fill="white"
-        />
-        <path
-          d="M105.82 172.001L83.8294 160.987C82.3492 160.252 80.869 158.05 80.869 156.214V129.413C80.869 127.578 82.1378 126.66 83.8294 127.578L104.974 138.224C105.82 138.592 106.455 138.224 106.455 137.307V98.3908C106.455 97.8401 105.609 97.2894 104.974 96.9223C104.129 96.5551 103.494 96.3716 103.494 96.9223V134.186L83.8294 124.274C80.6576 122.621 77.9087 124.274 77.9087 128.128V154.929C77.9087 158.6 80.4461 163.006 83.8294 164.658L105.82 175.672C106.666 175.855 106.666 172.368 105.82 172.001Z"
-          fill="#EBECF3"
-        />
-        <path
-          d="M106.243 44.0555C106.455 46.0748 106.878 47.9104 107.089 49.9296C107.512 51.7653 107.935 53.6009 108.358 55.253C108.358 55.6202 108.569 55.9873 108.781 56.3544C108.992 56.7216 109.415 56.9051 109.838 57.2723C111.953 58.3736 113.644 58.0065 115.97 57.823C117.873 57.6394 119.565 57.0887 121.256 56.3544C122.314 55.9873 123.159 55.4366 123.794 54.5188C124.217 53.7845 124.428 53.0502 124.428 52.1324C124.428 48.2775 122.525 44.7898 120.411 41.4856C119.776 40.5678 119.353 39.65 118.507 38.9157C116.816 37.2636 114.067 36.7129 111.318 36.3458C109.838 36.1622 107.935 35.9787 106.878 37.0801C106.032 37.8143 105.82 38.9157 105.82 39.8335C105.82 40.9349 106.032 42.587 106.243 44.0555Z"
-          fill="#FFCCAA"
-        />
-        <path
-          d="M104.974 38.5486C105.186 40.2007 105.82 41.8528 106.031 42.5871C106.031 42.7706 106.243 43.1378 106.454 43.3213C106.877 44.0556 107.512 44.7898 107.723 45.5241C107.934 46.4419 107.723 47.5433 108.569 48.2776C108.78 48.4612 108.992 48.6447 109.203 48.4612C109.415 48.4612 109.626 48.094 109.626 47.9105C109.837 47.3598 110.049 46.2584 110.683 46.0748C111.529 45.8912 112.586 46.9926 112.798 47.5433C113.221 48.4612 112.798 49.379 113.009 50.2968C113.221 52.316 113.855 53.9681 116.181 54.1517C117.661 54.3352 119.353 54.3352 120.833 53.9681C125.062 53.0503 127.811 50.1132 128.234 46.6255C128.868 42.9542 128.657 38.3651 126.119 35.0609C124.428 32.6745 120.833 31.9403 117.45 32.3074C114.489 32.6745 112.375 32.6745 108.146 34.6937C106.666 35.428 104.763 37.2637 104.974 38.5486Z"
-          fill="#2E2C34"
-        />
-        <path
-          d="M104.974 38.5485C105.186 40.2006 105.82 41.8527 106.031 42.5869C106.031 42.7705 106.243 43.1376 106.454 43.3212C110.472 40.7513 111.952 37.8142 108.357 34.6936C106.666 35.4279 104.763 37.2635 104.974 38.5485Z"
-          fill="#2E2C34"
-        />
-        <path
-          d="M65.01 62.2285L49.7856 69.9383L56.1291 73.0589L71.3536 65.5327L65.01 62.2285Z"
-          fill="#1E2A3D"
-        />
-        <path
-          d="M71.3532 65.533L56.1287 73.0591C56.1287 73.2427 56.1287 73.6098 56.1287 73.7934L71.3532 66.0837C71.5646 66.2672 71.5646 65.7165 71.3532 65.533Z"
-          fill="#1D293C"
-        />
-        <path
-          d="M63.7411 60.9439C63.5297 60.3932 63.3182 59.8425 63.5297 59.4753L67.7587 51.582L52.5342 59.2918L48.3052 67.1851C48.0937 67.5522 48.0937 68.1029 48.5166 68.6536C48.7281 69.0207 48.8163 69.271 49.0278 69.4546C49.2392 69.6382 49.6142 69.9653 49.7853 69.9386L65.0098 62.2288C64.7984 62.0453 64.5869 61.8617 64.3755 61.6781C64.164 61.4946 63.9526 61.311 63.7411 60.9439Z"
-          fill="#172030"
-        />
-        <path
-          d="M56.129 73.0592L49.7855 69.9386C49.3626 69.755 48.9397 69.2043 48.5168 68.6536C48.3053 68.1029 48.0939 67.5522 48.3053 67.1851L52.5343 59.2918V59.1082V51.3985C52.5343 51.2149 52.3229 50.8478 52.1114 50.8478C51.9 50.6642 51.6885 50.8478 51.6885 51.0313V58.7411L47.671 66.6344C47.4595 67.1851 47.4595 67.9194 47.8824 68.8372C48.3053 69.755 48.9397 70.4893 49.7855 70.8564L56.129 73.7837C56.3405 73.7837 56.3405 73.2428 56.129 73.0592Z"
-          fill="#1D293C"
-        />
-        <path
-          d="M67.7588 43.5051C67.7588 43.5051 67.7588 43.3215 67.5473 43.3215L67.3359 43.1379H67.1244L51.9 50.8477H52.1114C52.1114 50.8477 52.3229 50.8477 52.3229 51.0312L52.5343 51.2148V51.3984V59.1081V59.2917L67.7588 51.5819V51.3984V43.6886V43.5051Z"
-          fill="#B2B9BB"
-        />
-        <path
-          d="M93.5561 0.734131L93.7675 41.8528L24.4116 76.5467L23.7772 69.5712L23.3543 34.8773L93.5561 0.734131Z"
-          fill="#1E2A3D"
-        />
-        <path
-          d="M24.4116 76.5468L23.1429 75.9961L22.9314 34.6938L24.4116 35.4281V76.5468Z"
-          fill="#1D293C"
-        />
-        <path
-          d="M24.4116 35.4281L22.9315 34.6939L92.2874 0L93.5562 0.734262L24.4116 35.4281Z"
-          fill="#34455F"
-        />
-        <path
-          opacity="0.1"
-          d="M24.8346 35.6117L93.1333 1.46851V35.061L24.8346 69.2042V35.6117Z"
-          fill="#A39DAB"
-        />
-        <path
-          opacity="0.8"
-          d="M30.9666 35.2444C31.8124 34.8773 32.2353 35.9787 31.6009 36.8965C31.178 37.8144 30.3322 37.9979 30.3322 36.8965C29.9093 37.6308 29.0635 39.0993 29.0635 38.1815C29.0635 37.6308 29.275 36.5294 29.275 35.9787C29.275 35.7951 29.275 35.7951 29.4864 35.7951C29.4864 35.7951 29.4864 35.7951 29.4864 35.9787C29.275 36.713 29.275 37.4472 29.275 37.9979C29.4864 37.6308 29.6979 37.4472 29.9093 37.0801C30.3322 36.5294 30.5437 36.3458 30.5437 36.8965C30.5437 37.0801 30.7551 37.2637 30.9666 37.2637C31.8124 36.8965 31.8124 34.8773 30.9666 35.2444C30.9666 35.6116 30.7551 35.6116 30.9666 35.2444C30.7551 35.428 30.7551 35.428 30.9666 35.2444ZM35.407 34.5102C35.407 34.5102 33.2925 35.9787 33.2925 35.7951C33.0811 35.7951 35.407 34.3266 35.407 34.5102ZM35.407 33.7759L34.9841 33.9595C32.0238 35.9787 33.504 34.6937 35.407 33.7759C35.6185 33.7759 35.407 33.5923 35.407 33.7759ZM37.0987 35.428C36.8872 35.6116 36.6758 35.9787 36.6758 35.6116C36.6758 35.428 37.0987 34.143 37.0987 34.143C36.4643 35.2444 35.4071 35.428 36.0414 34.143C36.6758 32.6745 37.5216 32.491 37.5216 32.8581C37.5216 33.0416 37.5216 33.0416 37.3101 32.8581C36.6758 32.6745 35.6185 35.0609 36.2529 34.6937C36.6758 34.5102 37.0987 33.9595 37.3101 33.2252C37.3101 33.2252 37.5216 33.0417 37.5216 33.2252C37.5216 33.4088 36.8872 35.2444 36.8872 35.428C37.5216 34.8773 37.733 34.6937 37.0987 35.428ZM38.3674 31.5731C38.5788 31.7567 38.3674 32.3074 38.1559 32.491C37.5216 35.0609 39.4246 33.0417 39.4246 31.5731C39.4246 31.3896 39.4246 30.8389 39.6361 31.0224C39.8475 31.206 39.4246 32.6745 39.4246 33.0416C38.7903 34.3266 37.9445 34.3266 37.9445 33.7759C37.9445 33.5923 38.3674 31.5731 38.3674 31.5731ZM41.1162 32.8581C41.1162 32.6745 42.5964 28.8196 42.5964 29.0032C42.8078 29.1868 41.3277 32.491 41.1162 32.6745C41.3277 32.8581 41.3277 33.0416 41.1162 32.8581ZM43.8651 29.1868L43.6536 29.3703C43.4422 29.3703 44.288 28.2689 44.7109 28.4525C45.3453 28.6361 44.288 30.1046 43.8651 30.4717C43.6536 30.6553 43.4422 31.0224 43.2307 31.206C43.4422 31.206 44.7109 30.1046 44.7109 30.2882C44.7109 30.4717 43.4422 31.206 43.2307 31.206C43.0193 31.206 43.0193 31.206 43.0193 31.206C43.0193 31.0224 43.4422 30.6553 43.6536 30.4717C44.9223 29.0032 44.7109 28.2689 44.288 28.6361C44.0765 28.8196 43.8651 29.0032 43.8651 29.1868ZM47.4598 28.2689C47.6712 28.2689 47.6712 28.2689 47.6712 28.2689L46.614 29.0032C44.7109 30.2882 45.5567 29.3703 47.4598 28.2689ZM47.0369 28.0854C45.3452 29.1868 45.1338 29.0032 46.614 28.2689C47.8827 27.3511 48.517 27.1676 47.0369 28.0854ZM48.517 29.7375C48.517 29.7375 48.3056 30.1046 48.0941 30.1046C47.8827 30.2882 48.3056 28.8196 48.517 28.4525C48.3056 28.8196 47.6712 29.3703 47.4598 29.3703C46.8254 29.1868 48.0941 27.3511 48.517 27.1676C48.9399 26.984 49.1514 27.5347 48.7285 27.3511C48.517 27.3511 48.3056 27.5347 48.0941 27.9018C46.614 30.1046 48.3056 29.5539 48.7285 27.9018C48.7285 27.9018 48.9399 27.7183 48.9399 27.9018C48.9399 28.0854 48.3056 29.921 48.3056 30.1046C48.9399 29.1868 49.3628 29.0032 48.517 29.7375ZM50.4201 26.4333C50.6315 26.4333 49.7857 26.984 49.7857 26.984C49.3628 27.1676 49.5743 26.8004 49.9972 26.4333C50.6315 25.699 50.4201 25.3319 49.9972 25.699C49.7857 25.8826 49.5743 25.8826 49.9972 25.5155C50.4201 24.9648 50.843 25.3319 50.4201 26.0662C50.2086 26.2497 49.9972 26.4333 49.7857 26.8004C50.4201 26.2497 50.4201 26.4333 50.4201 26.4333ZM50.843 28.0854C50.6315 28.0854 52.3231 24.0469 52.3231 24.2305C52.746 24.0469 52.1117 25.3319 51.9002 25.699C51.4773 26.6169 51.4773 26.984 51.0544 27.9018C50.843 27.9018 50.843 28.0854 50.843 28.0854ZM53.3804 24.4141C53.1689 24.5976 52.9575 24.5976 53.3804 24.2305C54.4376 23.1291 55.4949 23.8634 52.9575 26.6169C53.1689 26.6169 54.4376 25.5155 54.4376 25.699C54.4376 25.8826 53.1689 26.6169 52.9575 26.6169C52.746 26.6169 52.746 26.6169 52.746 26.6169C52.746 26.4333 53.1689 26.0662 53.3804 25.8826C54.8605 23.8634 54.2262 23.4962 53.3804 24.4141ZM57.6094 21.8442V22.762C57.1865 21.8442 56.1293 22.5784 55.7064 23.4962C55.072 25.3319 56.5522 25.5155 57.6094 23.8634C57.398 24.2305 57.1865 24.4141 57.1865 24.5976C55.072 26.8004 54.0147 23.6798 56.5522 22.3949C56.9751 22.2113 57.398 22.2113 57.398 22.2113C57.398 22.0277 57.398 22.0277 57.6094 21.8442ZM60.3583 22.0277C60.3583 22.0277 58.2438 23.4962 58.2438 23.3127C58.0323 23.3127 60.3583 21.8442 60.3583 22.0277ZM60.3583 21.2935L59.9354 21.477C56.9751 23.4962 58.4552 22.0277 60.3583 21.2935C60.3583 21.2935 60.3583 21.1099 60.3583 21.2935ZM61.627 20.1921C60.9926 20.1921 60.1468 21.8442 60.1468 22.3949C60.1468 22.762 60.5697 22.762 61.2041 22.0277C61.8384 21.2935 61.627 21.8442 60.9926 22.3949C60.1468 23.1291 59.301 22.9456 59.9354 21.6606C60.3583 21.1099 60.9926 20.1921 61.627 20.1921C62.0499 20.1921 61.8384 19.8249 61.627 20.1921ZM62.8957 19.4578C63.1071 19.6414 62.4728 21.2935 62.6842 21.6606C62.8957 21.8442 64.1644 20.5592 63.9529 19.0907C63.9529 18.9071 63.9529 18.9071 63.9529 18.9071C63.9529 18.9071 64.1644 18.9071 64.1644 19.0907C64.1644 21.477 61.8384 23.4962 62.4728 20.7428C62.6842 19.8249 62.6842 19.4578 62.8957 19.4578C62.6842 19.4578 62.6842 19.2742 62.8957 19.4578ZM66.9133 18.1728C66.9133 17.9893 66.9133 17.9893 66.9133 18.1728C67.1247 17.9893 67.1247 17.9893 66.9133 18.1728C66.9133 18.3564 66.2789 18.7235 66.2789 18.7235H66.0675L66.4904 18.1728C66.7018 17.9893 66.9133 17.6222 66.9133 17.4386C66.9133 17.255 66.7018 17.255 66.4904 17.4386L66.2789 17.6222C66.0675 17.6222 66.2789 17.4386 66.2789 17.4386C66.4904 17.255 66.7018 17.0715 66.9133 17.255C67.1247 17.4386 66.9133 17.8057 66.9133 17.9893C66.7018 18.1729 66.4904 18.3564 66.2789 18.7235C66.9133 17.9893 66.9133 18.1728 66.9133 18.1728ZM67.3362 19.8249C67.1247 19.8249 68.6049 16.1536 68.8163 16.1536V15.9701C68.8163 15.9701 69.0278 15.9701 69.0278 16.1536C69.0278 16.3372 68.6049 17.255 68.6049 17.4386C68.182 18.3564 68.182 18.7235 67.7591 19.6414C67.5476 19.6414 67.5476 19.8249 67.3362 19.8249ZM70.085 16.1536C69.8736 16.1536 69.8736 16.1536 70.085 16.1536C69.8736 16.3372 69.6621 16.3372 70.085 15.9701C70.2965 15.7865 70.7194 15.4194 70.9308 15.6029C71.1423 15.7865 71.1423 16.1536 70.9308 16.5208C70.7194 17.0715 69.8736 17.9893 69.4507 18.54C69.4507 18.54 69.8736 18.3564 70.085 18.1728C70.9308 17.6222 70.9308 17.6222 70.9308 17.6222C70.9308 17.8057 69.6621 18.7235 69.2392 18.7235C69.2392 18.7235 69.2392 18.54 69.4507 18.54C69.4507 18.3564 69.6621 18.1728 69.8736 18.1728C71.5652 15.6029 70.7194 15.0522 70.085 16.1536ZM29.6979 41.3021C29.6979 41.1185 29.6979 41.3021 29.6979 41.3021C30.1208 41.4857 29.275 43.872 29.275 43.872C29.0635 44.0556 29.275 42.9542 29.4864 42.5871C29.6979 41.4857 29.6979 41.3021 29.6979 41.3021ZM30.9666 43.6884L31.178 41.3021C31.178 41.1185 31.3895 41.1185 31.3895 41.3021C31.3895 41.4857 31.178 43.872 30.9666 43.6884ZM34.3498 40.3843C34.3498 40.3843 32.2353 41.8528 32.2353 41.6692C31.8124 41.6692 34.3498 40.2007 34.3498 40.3843ZM34.3498 39.65L33.9269 39.8336C30.9666 41.8528 32.2353 40.5678 34.3498 39.65C34.3498 39.65 34.3498 39.4664 34.3498 39.65ZM35.1956 38.5486C35.1956 38.5486 35.4071 38.365 35.1956 38.5486C35.4071 38.7322 34.7727 41.3021 34.7727 41.1185C34.5612 41.1185 35.1956 38.365 35.1956 38.5486ZM36.8872 38.5486L36.6758 38.7322C36.4643 38.7322 37.3101 37.6308 37.733 37.8144C38.5788 38.1815 36.8872 40.0171 36.2529 40.5678C36.4643 40.3843 37.5216 39.8336 37.733 39.65C38.3674 39.4664 36.0414 41.1185 36.2529 40.5678C36.2529 40.3843 37.0987 39.4664 37.3101 39.2829C38.1559 38.1815 37.733 37.8144 37.5216 37.8144C37.3101 38.365 37.0987 38.5486 36.8872 38.5486ZM39.8475 37.6308C39.8475 37.6308 37.733 39.0993 37.733 38.9157C37.5216 38.9157 39.8475 37.4472 39.8475 37.6308ZM39.8475 36.8965L39.4246 37.0801C36.4643 39.0993 37.733 37.8144 39.8475 36.8965C40.059 36.713 39.8475 36.713 39.8475 36.8965ZM40.6933 35.7951C40.6933 35.6116 40.9048 35.6116 40.6933 35.7951C41.1162 35.9787 40.2704 38.365 40.2704 38.365C40.059 38.5486 40.2704 37.4472 40.4819 37.0801C40.6933 35.9787 40.6933 35.7951 40.6933 35.7951ZM41.5391 38.1815C41.962 37.8144 41.7506 37.6308 41.7506 37.6308C41.5391 37.6308 41.7506 37.2637 41.962 37.2637C42.1735 37.2637 41.962 37.8144 41.7506 37.9979C41.3277 38.5486 41.1162 38.7322 41.5391 38.1815ZM44.4994 33.9595C44.7109 34.143 44.4994 34.6937 44.288 34.8773C43.6536 37.4472 45.5567 35.428 45.5567 33.9595C45.5567 33.7759 45.5567 33.2252 45.7682 33.4088C46.1911 33.5923 45.5567 35.7951 44.4994 36.3458C44.0765 36.5294 43.8651 36.3458 43.8651 35.9787C43.8651 35.9787 44.288 33.9595 44.4994 33.9595C44.288 33.9595 44.288 33.9595 44.4994 33.9595ZM47.4598 35.2444C47.4598 35.2444 47.6712 34.143 47.6712 33.0416C47.6712 32.8581 47.8827 32.8581 47.8827 33.0416C47.8827 35.2444 47.4598 35.6116 47.4598 35.2444ZM49.7857 32.491C49.7857 32.6745 49.5743 33.7759 49.5743 33.4088C49.5743 33.4088 49.5743 32.6745 49.7857 32.491C49.5743 32.6745 48.9399 33.0416 48.7285 33.0416C48.3056 32.8581 49.3628 32.491 49.7857 32.491C49.5743 32.3074 49.7857 31.5731 49.7857 31.5731C49.7857 31.5731 49.9972 31.5731 49.9972 31.7567C49.9972 31.7567 49.9972 32.1238 49.7857 32.491C49.9972 32.3074 50.6315 31.9403 50.843 31.9403C51.0544 31.5731 51.2659 31.7567 49.7857 32.491ZM51.9002 30.2882C52.1117 30.4717 51.9002 31.0224 51.6888 31.206C51.0544 33.7759 53.1689 31.7567 52.9575 29.921C52.9575 29.7375 52.9575 29.7375 52.9575 29.7375C52.9575 29.7375 53.1689 29.7375 53.1689 29.921C53.1689 32.491 50.843 34.143 51.4773 31.5731C51.6888 30.6553 51.6888 30.2882 51.9002 30.2882C51.6888 30.2882 51.6888 30.1046 51.9002 30.2882ZM55.4949 29.3703L55.2835 29.5539C55.072 29.5539 55.9178 28.4525 56.3407 28.6361C57.1865 29.0032 55.4949 30.8389 54.8605 31.3896C55.072 31.206 56.1293 30.6553 56.3407 30.4717C56.9751 30.1046 55.7064 31.206 55.072 31.3896C54.8605 31.3896 54.8605 31.3896 54.8605 31.3896C54.8605 31.206 55.2834 30.8389 55.4949 30.6553C55.9178 30.2882 56.7636 29.0032 56.3407 28.6361H56.1293C55.7064 29.0032 55.4949 29.1868 55.4949 29.3703ZM56.9751 29.0032C59.7239 27.3511 57.6094 29.0032 56.1293 29.7375C55.9178 29.5539 56.7636 29.1868 56.9751 29.0032ZM58.2438 27.5347C58.2438 27.5347 57.6094 27.9018 57.398 28.0854C57.1865 28.0854 56.1293 28.8196 56.1293 28.8196H55.9178C56.1293 28.6361 58.0323 27.7183 58.2438 27.5347C58.4552 27.5347 58.4552 27.5347 58.2438 27.5347ZM59.301 26.6169C59.5125 26.8004 59.301 27.3511 59.0896 27.5347C58.4552 30.1046 60.3583 28.0854 60.3583 26.2497C60.3583 26.0662 60.3583 26.0662 60.3583 26.0662C60.7812 25.8826 60.5697 28.4525 59.301 29.0032C58.8781 29.1868 58.6667 29.0032 58.6667 28.6361C58.6667 28.4525 59.0896 26.6169 59.301 26.6169C59.0896 26.6169 59.0896 26.4333 59.301 26.6169ZM61.627 28.2689C61.627 28.2689 62.0499 27.7183 62.0499 27.5347C62.0499 27.3511 62.0499 27.3511 61.8384 27.3511C61.627 27.3511 61.627 27.3511 61.8384 27.1676C62.0499 26.8004 62.6842 26.984 62.0499 27.7183C62.0499 27.9018 61.627 28.4525 61.627 28.2689ZM63.7415 26.4333L64.1644 24.2305C63.7415 24.2305 64.7987 23.4962 65.4331 23.4962C66.0675 23.4962 65.4331 24.4141 65.0102 24.7812C64.3758 25.3319 63.9529 25.3319 65.2216 25.699C65.2216 25.699 65.4331 25.699 65.4331 25.8826C65.4331 26.2497 64.1644 25.8826 63.9529 25.699C63.9529 25.8826 63.7415 26.2497 63.7415 26.6169C63.9529 26.4333 63.9529 26.6169 63.7415 26.4333ZM64.3758 25.1483C64.5873 24.7812 65.6446 24.2305 65.6446 23.6798C65.6446 23.3127 65.0102 23.6798 64.5873 24.0469C64.5873 24.2305 64.3758 25.1483 64.3758 25.1483ZM67.1247 25.3319C67.1247 25.3319 67.3362 24.2305 67.3362 23.1291C67.3362 22.9456 67.5476 22.9456 67.5476 23.1291C67.5476 25.5155 67.1247 25.8826 67.1247 25.3319ZM69.4507 22.5784C69.4507 22.5784 69.2392 23.4962 69.2392 23.6798C69.0278 23.6798 69.2392 22.9456 69.4507 22.5784C69.2392 22.762 68.3934 23.1291 68.3934 23.1291C68.3934 23.1291 68.182 22.9456 68.3934 22.9456C68.8163 22.762 68.8163 22.762 69.4507 22.5784C69.2392 22.3949 69.4507 21.6606 69.4507 21.6606C69.6621 21.477 69.4507 22.2113 69.4507 22.3949C69.6621 22.2113 70.2965 21.8442 70.5079 21.8442C70.7194 21.6606 70.7194 21.8442 69.4507 22.5784ZM70.9308 22.9456C70.7194 22.9456 71.1423 22.0277 71.1423 21.8442C71.3537 21.2935 71.3537 20.9263 71.3537 20.7428C70.9308 20.7428 71.9881 20.0085 72.6224 20.0085C72.8339 20.0085 72.8339 20.0085 72.8339 20.1921C72.8339 20.5592 72.411 20.9263 72.1995 21.1099C71.5652 21.6606 71.1423 21.6606 72.411 22.0277C72.411 22.0277 72.6224 22.0277 72.6224 22.2113C72.6224 22.3949 71.9881 22.3949 71.7766 22.2113C71.3537 22.2113 71.1423 22.0277 71.1423 21.8442C71.1423 22.0277 70.9308 22.2113 70.9308 22.5784V22.762C70.9308 22.762 70.9308 23.1291 70.9308 22.9456ZM71.3537 21.6606C71.7766 21.2935 71.9881 21.1099 72.411 20.5592C72.8339 20.0085 72.411 19.8249 71.5652 20.5592C71.5652 20.7428 71.3537 21.477 71.3537 21.6606ZM74.7369 19.8249L74.5255 20.0085C74.314 20.0085 75.1598 18.9071 75.5827 19.0907C76.2171 19.2742 75.1599 20.7428 74.7369 21.1099C74.5255 21.2935 74.314 21.6606 74.1026 21.8442C74.1026 21.8442 74.5255 21.6606 74.7369 21.477C76.0057 20.7428 75.7942 20.9263 74.9484 21.477C74.7369 21.6606 73.8911 22.2113 73.8911 22.0277C73.8911 21.8442 74.7369 20.9263 74.9484 20.7428C75.7942 19.6414 75.3713 19.2742 75.1598 19.2742C74.9484 19.4578 74.7369 19.6414 74.7369 19.8249ZM76.2171 19.4578C78.1202 18.3564 78.1202 18.54 76.0057 19.8249C75.1599 20.1921 74.9484 20.1921 76.2171 19.4578ZM77.4858 17.9893L77.0629 18.1728C74.1026 20.1921 75.3713 18.9071 77.4858 17.9893C77.6973 17.9893 77.6973 17.8057 77.4858 17.9893ZM77.9087 19.4578L78.3316 17.255C77.9087 17.255 78.966 16.5208 79.6003 16.5208C80.0232 16.5208 79.8118 17.0715 79.6003 17.255C78.966 18.1728 77.9087 17.9893 79.1774 18.3564C80.4461 18.7235 78.5431 18.7235 78.3316 18.1728C78.3316 18.3564 78.1202 18.7235 78.1202 19.0907C78.1202 19.2742 77.9087 19.4578 77.9087 19.4578ZM78.3316 18.1728C78.5431 17.8057 79.6003 17.255 79.6003 16.7043C79.6003 16.3372 78.966 16.7043 78.5431 17.0715C78.5431 17.255 78.3316 17.9893 78.3316 18.1728ZM31.178 47.9104L31.6009 48.4611C31.6009 48.6447 31.3895 48.6447 31.178 48.094C29.4864 50.1132 29.9093 46.4419 31.3895 45.7077C31.8124 45.3405 32.4467 46.2584 31.178 47.9104ZM30.9666 47.9104C30.7551 47.5433 30.5437 46.8091 30.7551 46.4419C29.6979 47.7269 29.9093 49.1954 30.9666 47.9104ZM31.178 47.7269C31.3895 47.3598 32.0238 46.2584 31.6009 45.8912C31.178 45.5241 30.5437 46.4419 31.178 47.7269ZM34.7727 45.3405C34.7727 45.3405 32.6582 46.8091 32.6582 46.6255C32.4467 46.6255 34.7727 45.157 34.7727 45.3405ZM34.7727 44.6063L34.3498 44.7898C31.3895 46.8091 32.6582 45.5241 34.7727 44.6063C34.9841 44.6063 34.9841 44.4227 34.7727 44.6063ZM35.83 43.5049C35.83 43.3213 35.83 43.3213 35.83 43.5049C36.2529 43.6884 35.407 46.0748 35.407 46.0748C35.1956 46.2584 35.4071 45.157 35.6185 44.7898C35.83 43.6884 35.83 43.5049 35.83 43.5049ZM37.0987 43.1377C36.8872 43.1377 36.8872 43.1377 37.3101 42.7706C37.5216 42.5871 37.9445 42.2199 38.1559 42.4035C39.0017 42.7706 37.3101 44.6063 36.6758 45.157C36.8872 45.157 38.1559 44.0556 38.1559 44.2391C38.1559 44.4227 36.8872 45.157 36.6758 45.157C36.4643 45.157 36.4643 45.157 36.4643 45.157C36.4643 44.9734 36.8872 44.6063 37.0987 44.4227C38.3674 42.9542 38.1559 42.2199 37.733 42.5871C37.5216 42.7706 37.3101 43.1377 37.0987 43.1377ZM39.0017 44.2391L39.4246 42.0364C39.2132 42.0364 39.4246 41.8528 39.6361 41.6692C40.9048 40.7514 41.7506 41.3021 39.8475 42.7706C39.6361 42.9542 39.4246 42.9542 40.2704 43.1377C41.5391 43.5049 39.6361 43.5049 39.4246 42.9542C39.4246 43.1377 39.2132 44.2391 39.0017 44.2391ZM39.6361 42.7706L39.8475 42.5871C41.3277 41.3021 40.9048 40.935 39.8475 41.8528L39.6361 42.7706ZM41.962 41.1185C41.962 41.4857 41.3277 42.7706 41.7506 42.5871C42.1735 42.4035 42.3849 42.0364 42.3849 42.0364C42.8078 41.8528 41.7506 42.9542 41.5391 42.7706C41.3277 42.7706 41.7506 41.4857 41.7506 41.1185C41.3277 41.6692 41.1162 41.4857 41.962 41.1185C41.7506 40.7514 41.7506 40.7514 41.962 40.5678V40.3843C42.1735 40.2007 42.1735 40.7514 41.962 40.935C42.8078 40.3843 43.2307 40.3843 41.962 41.1185ZM79.8118 12.8495H80.4461C80.4461 13.033 79.8118 13.2166 79.3889 13.033C78.3316 12.6659 79.3889 10.6467 80.0232 9.72884C80.2347 9.54528 80.869 8.62745 80.869 8.99458C80.869 8.99458 80.0232 10.096 79.8118 10.2795C79.3889 11.1974 78.7545 12.6659 79.8118 12.8495ZM81.0805 12.1152L81.5034 9.91241C81.2919 9.72884 81.9263 9.17814 82.1377 9.17814C82.5606 8.99458 82.9835 8.99458 82.9835 9.36171C82.9835 9.72884 82.5606 10.096 82.3492 10.2795C81.7148 10.8302 81.2919 10.8302 82.5606 11.0138C82.9835 11.1974 82.7721 11.3809 82.1377 11.1974C81.7148 11.1974 81.5034 11.0138 81.5034 10.8302C81.5034 11.0138 81.2919 12.1152 81.0805 12.1152ZM81.5034 10.8302C81.7148 10.4631 82.7721 9.91241 82.7721 9.36171C82.7721 8.99458 82.1377 9.36171 81.7148 9.72884C81.7148 9.72884 81.5034 10.6467 81.5034 10.8302ZM85.3095 8.99458C85.521 8.99458 85.521 8.99458 85.521 8.99458C85.521 8.99458 84.8866 9.54527 84.6751 9.54527C82.1377 11.0138 83.4064 9.91241 85.3095 8.99458ZM85.521 8.07675L85.098 8.26032C82.1377 10.4631 83.6179 8.99458 85.521 8.07675C85.7324 8.07675 85.7324 7.89319 85.521 8.07675ZM86.7897 7.15892C87.2126 6.42466 87.6355 6.42466 87.6355 7.34249C87.6355 7.70962 87.6355 8.07675 87.424 8.44388C86.5782 10.096 85.5209 8.81101 86.7897 7.15892ZM87.424 7.34249C87.424 6.60823 87.2126 6.60823 86.7897 7.15892C86.5782 7.34249 86.5782 7.52606 86.5782 7.89319C85.9439 9.72884 87.424 8.99458 87.424 7.34249ZM88.0584 8.99458C87.8469 9.17814 87.6355 9.17814 87.6355 9.17814L88.2698 8.44388C88.9042 7.70962 89.1156 7.34249 89.1156 6.97536C89.3271 6.60823 89.3271 6.42466 89.3271 6.2411C89.3271 6.05753 89.3271 5.87397 89.3271 5.87397C89.3271 5.6904 89.1156 5.6904 89.1156 5.6904H88.9042H88.6927C88.6927 5.6904 88.6927 5.6904 88.9042 5.50684C89.1156 5.50684 89.3271 5.50684 89.5385 5.50684C89.75 5.6904 89.75 5.87397 89.75 6.2411C89.75 6.79179 89.5385 7.34249 89.3271 7.70962C89.1156 7.89319 89.1156 8.07675 88.9042 8.44388C88.4813 8.26032 88.2698 8.62745 88.0584 8.99458ZM80.2347 22.2113C80.0232 22.3949 79.6003 22.5784 79.1774 22.762C79.1774 22.9456 78.7545 24.2305 78.7545 24.0469C78.7545 24.0469 78.7545 23.4962 78.966 23.3127C79.1774 22.3949 79.3889 21.477 79.3889 21.477C79.6003 21.2935 80.869 20.5592 80.869 20.7428C80.869 20.7428 80.6576 20.9263 80.4461 20.9263C80.2347 21.1099 80.0232 21.1099 79.8118 21.2935C79.6003 21.2935 79.6003 21.477 79.3889 21.6606C79.3889 21.8442 79.1774 22.3949 79.1774 22.5784C79.8118 22.3949 80.6576 21.8442 80.2347 22.2113ZM82.3492 21.8442C81.5034 23.4962 80.0232 23.3127 81.7148 21.1099C82.5606 20.0085 82.9835 20.5592 82.7721 21.1099C82.7721 21.1099 82.5606 21.6606 82.3492 21.8442C82.5606 21.8442 82.5606 22.2113 82.5606 22.2113C82.5606 22.2113 82.5606 22.2113 82.3492 22.2113C82.3492 22.3949 82.3492 22.0277 82.3492 21.8442ZM82.5606 21.1099C82.5606 20.9263 82.5606 20.9263 82.5606 20.7428C82.3492 20.5592 81.9263 20.9263 81.9263 21.2935C80.2347 23.3127 81.9263 23.1291 82.5606 21.1099ZM84.2522 20.5592C85.7324 19.6414 86.1553 19.6414 84.8866 20.3756C84.8866 20.3756 83.6179 21.2935 83.4064 21.1099C83.195 21.1099 84.0408 20.7428 84.2522 20.5592ZM85.521 19.0907C85.3095 19.2742 83.6179 20.1921 83.4064 20.3756H83.195C83.4064 20.1921 85.3095 19.0907 85.521 19.0907C85.7324 19.0907 85.7324 18.9071 85.521 19.0907ZM86.5782 17.9893C86.5782 17.8057 86.5782 17.9893 86.5782 17.9893C86.7897 18.1729 86.5782 19.2742 86.5782 19.4578C86.5782 19.4578 86.3668 20.5592 86.1553 20.5592C85.9439 20.7428 86.1553 19.6414 86.3668 19.2742C86.5782 18.1729 86.5782 17.9893 86.5782 17.9893ZM87.2126 20.0085C87.0011 20.0085 87.0011 19.8249 87.0011 19.8249C87.0011 19.8249 87.2126 19.8249 87.424 19.8249C87.424 19.8249 87.424 19.8249 87.424 19.6414C87.424 19.6414 87.8469 17.9893 87.8469 17.6222C87.8469 17.6222 87.6355 17.8057 87.6355 17.6222C87.424 17.6222 87.6355 17.4386 87.8469 17.4386C89.3271 16.3372 90.1729 17.0715 88.4813 18.3564C88.9042 18.1729 89.1156 18.1728 89.3271 18.1728C89.9614 17.9893 88.0584 20.0085 87.2126 20.0085ZM88.0584 18.54C87.8469 18.54 87.8469 18.54 88.0584 18.54C87.8469 18.7235 87.6355 19.6414 87.6355 19.8249C88.9042 18.9071 89.75 17.6222 88.0584 18.54ZM88.9042 17.0715C88.6927 17.0715 88.2698 17.255 88.0584 17.4386C88.2698 17.4386 88.0584 18.1729 87.8469 18.3564C87.8469 18.3564 88.0584 18.1728 88.2698 18.1728C88.4813 17.9893 88.6927 17.8057 88.9042 17.6222C88.9042 17.4386 89.1156 17.0715 88.9042 17.0715ZM28.0063 53.9681C27.7948 54.1517 27.3719 54.3352 26.949 54.5188C26.949 54.7024 26.5261 55.9873 26.5261 55.8038L27.1605 53.2338C27.3719 53.0503 28.6406 52.316 28.6406 52.4996L27.7948 53.0503C27.5834 53.0503 27.5834 53.2338 27.3719 53.4174C27.3719 53.601 27.1605 54.1517 27.1605 54.3352C27.5834 54.1517 28.4292 53.601 28.0063 53.9681ZM29.6979 53.4174C31.6009 52.316 31.6009 52.4996 29.4864 53.7845C28.6406 54.1517 28.4292 54.1517 29.6979 53.4174ZM30.9666 51.9489L30.5437 52.1325C27.7948 54.1517 28.8521 52.8667 30.9666 51.9489C31.178 51.9489 31.178 51.7653 30.9666 51.9489ZM31.3895 53.4174C31.178 53.4174 31.178 53.2338 31.178 53.2338C31.178 53.2338 31.3895 53.2338 31.6009 53.2338C31.3895 53.0503 31.6009 52.8667 31.6009 52.6832C32.0238 51.3982 32.0238 51.0311 32.0238 51.0311C31.8124 51.2146 31.6009 51.0311 31.8124 50.8475C33.2925 49.7461 34.1383 50.4804 32.4467 51.7653C34.7727 50.4804 32.6582 53.4174 31.3895 53.4174ZM32.2353 51.9489C32.0238 52.1325 31.8124 53.0503 31.8124 53.2338C33.2925 52.4996 34.1383 51.0311 32.2353 51.9489ZM32.2353 50.8475C32.4467 50.8475 32.2353 51.5818 32.0238 51.7653C32.2353 51.5818 32.6582 51.2146 32.8696 51.0311C33.504 50.6639 33.2925 50.2968 32.2353 50.8475ZM35.407 52.316C34.5612 53.2338 34.9841 52.316 35.407 51.0311C34.7727 52.1325 33.7154 52.316 34.3498 51.0311C34.5612 50.6639 34.9841 49.9297 35.407 49.7461C35.83 49.5625 36.0414 50.1132 35.6185 49.9297C34.9841 49.7461 33.9269 52.1325 34.5612 51.7653C35.1956 51.3982 35.6185 50.2968 35.6185 50.2968C35.83 50.1132 35.6185 50.8475 35.407 51.2146C35.1956 51.7653 34.9841 52.316 34.9841 52.4996C35.83 51.7653 36.0414 51.5818 35.407 52.316ZM37.9445 48.4611C38.3674 48.8283 38.1559 50.2968 37.0987 50.8475C36.2529 51.2146 36.2529 50.6639 36.2529 49.9297C36.2529 49.5625 36.2529 49.379 36.2529 49.379L36.6758 49.1954C36.6758 49.1954 36.6758 50.1132 36.6758 50.2968C36.6758 51.0311 37.733 50.6639 37.733 49.379C37.733 49.0118 37.733 48.8283 37.5216 48.6447L37.9445 48.4611ZM39.8475 47.7269C39.6361 47.7269 39.4246 47.9104 39.2132 48.094C38.7903 48.4611 39.0017 48.6447 39.4246 48.6447C40.2704 48.6447 39.8475 49.379 39.2132 49.7461C39.0017 49.9297 38.3674 49.9297 38.3674 49.7461C38.3674 49.7461 38.3674 49.5625 38.5788 49.5625C39.2132 49.7461 40.2704 48.6447 39.2132 48.8283C39.0017 48.8283 38.7903 48.8283 38.7903 48.6447C38.7903 48.2776 39.4246 47.7269 39.6361 47.5433C40.059 47.3598 40.059 47.9104 39.8475 47.7269ZM40.9048 46.9926C41.1162 46.9926 39.8475 51.2146 40.4819 47.9104C40.4819 47.7269 40.6933 46.8091 40.9048 46.9926ZM41.1162 46.4419C40.9048 46.4419 40.6933 46.4419 40.9048 46.2584C41.1162 45.8912 41.3277 46.2584 41.1162 46.4419ZM43.2307 45.8912C43.4422 46.0748 43.2307 46.6255 43.2307 46.8091C42.8078 48.094 42.5964 47.9104 42.8078 47.5433C42.8078 47.1762 43.2307 46.4419 43.2307 46.0748C43.2307 45.3405 42.3849 46.8091 42.1735 47.3598C40.9048 49.5625 41.962 47.5433 41.962 46.4419C41.962 46.2584 42.1735 46.2584 42.1735 46.4419C42.1735 46.8091 41.962 47.5433 41.7506 47.7269C41.7506 47.5433 42.5964 45.7077 43.2307 45.8912ZM47.0369 43.872C47.0369 44.0556 47.0369 44.7898 47.0369 45.157C47.0369 45.3405 47.0369 45.5241 47.2483 45.3405V45.5241C47.0369 45.7077 46.8254 45.7077 46.8254 45.3405C46.4025 46.2584 45.1338 46.6255 45.1338 45.5241C45.1338 44.4227 46.4025 43.6884 46.8254 44.0556C46.8254 43.872 46.8254 43.872 46.8254 43.6884L47.0369 43.872ZM46.614 45.3405C46.614 45.157 46.614 44.9734 46.614 44.7898C46.4025 44.2391 45.5567 44.7898 45.5567 45.7077C45.3453 46.4419 46.4025 46.0748 46.614 45.3405ZM32.0238 56.9052C33.504 55.9873 33.7154 55.9873 32.6582 56.7216C30.7551 57.823 30.7551 57.6394 32.0238 56.9052ZM32.8696 55.8038C29.9093 57.6394 31.6009 56.1709 33.2925 55.4366C33.504 55.2531 33.7154 55.2531 32.8696 55.8038ZM33.7154 56.9052C33.504 56.9052 33.504 56.7216 33.504 56.7216C33.504 56.7216 33.7154 56.7216 33.9269 56.7216C33.7154 56.538 33.9269 56.3545 33.9269 56.1709C34.3498 54.8859 34.3498 54.5188 34.3498 54.5188C34.1383 54.7024 33.9269 54.5188 34.1383 54.3352C35.6185 53.2338 36.4643 53.9681 34.7727 55.2531C37.0987 53.9681 34.9841 56.9052 33.7154 56.9052ZM34.5612 55.4366C34.3498 55.6202 34.1383 56.538 34.1383 56.7216C35.6185 55.9873 36.4643 54.5188 34.5612 55.4366ZM34.5612 54.3352C34.7727 54.3352 34.5612 55.0695 34.3498 55.2531C34.5612 55.0695 34.9842 54.7024 35.1956 54.5188C35.83 54.1517 35.6185 53.7845 34.5612 54.3352ZM36.2529 55.2531C36.2529 55.0695 36.2529 55.2531 36.2529 55.2531C37.0987 55.4366 38.5788 54.3352 37.3101 53.9681C36.8872 53.7845 36.6758 53.601 36.8872 53.4174C37.0987 53.0503 38.1559 52.316 38.1559 52.4996L37.5216 53.0503C36.8872 53.601 37.0987 53.601 37.5216 53.7845C37.733 53.7845 37.9445 53.9681 37.9445 54.1517C38.1559 55.0695 36.2529 55.9873 36.2529 55.2531ZM40.2704 53.7845C39.6361 54.5188 38.5788 54.7024 38.5788 53.7845C38.5788 52.8667 39.6361 52.1325 40.2704 51.9489V52.316C38.7903 52.316 38.5788 55.0695 40.2704 53.601V53.7845ZM40.6933 53.4174C40.059 52.8667 41.5391 50.8475 41.962 51.2146C42.5964 51.7653 41.1162 53.7845 40.6933 53.4174ZM40.9048 53.2338C41.5391 52.8667 42.1735 51.2146 41.5391 51.3982C40.6933 51.9489 40.4819 53.2338 40.9048 53.2338ZM43.8651 50.4804C43.6536 50.4804 43.4422 50.6639 43.2307 50.8475C42.8078 51.2146 43.0193 51.3982 43.4422 51.3982C44.288 51.3982 43.8651 52.1325 43.2307 52.4996C43.0193 52.6832 42.3849 52.6832 42.3849 52.4996C42.3849 52.4996 42.3849 52.316 42.5964 52.316C43.2307 52.316 44.0765 51.3982 43.2307 51.3982C43.0193 51.3982 42.8078 51.3982 42.8078 51.2146C42.8078 50.8475 43.4422 50.2968 43.6536 50.1132C44.0765 49.9297 44.0765 50.4804 43.8651 50.4804ZM47.6712 48.2776C47.6712 48.4611 47.6712 49.1954 47.6712 49.5625C47.6712 49.7461 47.6712 49.9297 47.8827 49.7461V49.9297C47.6712 50.1132 47.4598 50.1132 47.4598 49.7461C47.2483 50.1132 47.0369 50.2968 46.8254 50.4804C45.5567 51.0311 45.7682 49.0118 46.8254 48.4611C47.0369 48.2776 47.4598 48.2776 47.4598 48.4611C47.4598 48.2776 47.4598 48.2776 47.4598 48.094L47.6712 48.2776ZM46.614 50.4804C46.8254 50.2968 47.0369 49.9297 47.2483 49.7461C47.4598 49.1954 47.2483 48.8283 46.614 49.0118C45.7682 49.379 45.7682 50.8475 46.614 50.4804ZM57.6094 42.5871C58.2438 42.4035 58.8781 42.5871 58.8781 43.1377C58.8781 43.6884 58.6667 44.0556 58.2438 44.4227C58.0323 44.6063 57.8209 44.7898 57.6094 44.9734C57.398 45.157 57.6094 45.5241 57.398 45.7077C57.398 45.7077 57.398 45.7077 57.398 45.5241C57.398 45.157 57.398 44.9734 56.9751 45.157C56.7636 45.3405 56.5522 45.3405 56.1293 45.157C55.9178 44.9734 55.9178 44.7898 55.9178 44.4227C55.9178 44.0556 56.1293 43.6884 56.3407 43.3213C56.5522 43.1377 56.7636 42.9542 56.9751 42.7706C57.1865 42.7706 57.398 42.5871 57.6094 42.5871ZM57.1865 43.3213C56.9751 43.3213 56.7636 43.5049 56.5522 43.6884C56.3407 43.872 56.3407 44.2391 56.3407 44.4227C56.3407 45.157 57.1865 44.9734 57.1865 44.4227C57.1865 44.2391 57.398 43.5049 57.1865 43.3213ZM57.8209 44.4227C58.0323 44.4227 58.2438 44.2391 58.4552 43.872C58.4552 43.6884 58.6667 43.1377 58.4552 42.9542C58.2438 42.7706 57.8209 42.7706 57.6094 42.9542C57.398 43.1377 57.398 43.872 57.6094 44.0556C57.6094 44.2391 57.6094 44.4227 57.8209 44.4227ZM29.6979 56.538C30.3322 56.3545 30.9666 56.538 30.9666 57.0887C30.9666 57.6394 30.7551 58.0065 30.3322 58.3737C30.1208 58.5572 29.9093 58.7408 29.6979 58.9244C29.4864 59.1079 29.6979 59.4751 29.4864 59.6586C29.4864 59.6586 29.4864 59.6586 29.4864 59.4751C29.4864 59.1079 29.4864 58.9244 29.0635 59.1079C28.8521 59.2915 28.6406 59.2915 28.2177 59.1079C28.0063 58.9244 28.0063 58.5572 28.0063 58.3737C28.0063 58.0065 28.2177 57.6394 28.4292 57.2723C28.6406 57.0887 28.8521 56.9052 29.0635 56.7216C29.275 56.7216 29.4864 56.7216 29.6979 56.538ZM29.0635 57.4558C28.8521 57.4558 28.6406 57.6394 28.4292 57.823C28.4292 58.0065 28.4292 58.1901 28.4292 58.5572C28.4292 59.2915 29.275 59.1079 29.275 58.5572C29.275 58.1901 29.275 57.4558 29.0635 57.4558C29.275 57.4558 29.0635 57.4558 29.0635 57.4558ZM29.6979 58.3737C29.9093 58.3737 30.1208 58.1901 30.3322 57.823C30.3322 57.6394 30.5437 57.0887 30.3322 56.9052C30.1208 56.7216 29.6979 56.7216 29.4864 56.9052C29.275 57.0887 29.275 57.823 29.4864 58.0065C29.4864 58.3737 29.6979 58.3737 29.6979 58.3737ZM60.1468 42.7706C61.8384 41.8528 61.627 42.0364 61.2041 42.4035C58.6667 43.872 59.0896 43.5049 60.1468 42.7706ZM59.5125 42.5871C59.301 42.5871 59.5125 42.5871 59.5125 42.5871C59.301 42.4035 61.2041 41.3021 61.4155 41.3021C62.2613 40.935 60.3583 42.2199 59.5125 42.5871ZM62.2613 42.4035C62.8957 42.0364 63.53 41.4857 63.53 41.4857C64.1644 41.1185 61.8384 43.1377 61.8384 42.4035C61.8384 42.2199 62.2613 40.3843 62.2613 40.2007C62.2613 40.0171 62.2613 40.0171 62.2613 40.0171C63.1071 40.5678 61.8384 42.7706 62.2613 42.4035ZM64.7987 39.0993C64.7987 38.9157 64.7987 39.0993 64.7987 39.0993C65.0102 39.0993 65.0102 39.0993 64.7987 39.0993C65.0102 40.3843 63.7415 43.5049 64.5873 40.5678C64.7987 39.2829 64.7987 39.0993 64.7987 39.0993ZM77.9087 29.0032C77.9087 29.1868 77.0629 29.5539 76.8515 29.7375C76.8515 29.921 76.8515 30.1046 76.64 30.2882C76.4286 30.8389 76.2171 31.206 77.6973 30.1046C78.1202 29.7375 78.1202 29.921 77.4858 30.4717C75.1599 32.1238 76.64 29.921 76.64 28.8196C76.64 28.6361 77.0629 28.2689 77.4858 28.2689C78.966 27.5347 78.3316 28.0854 76.8515 28.8196C76.8515 29.0032 76.64 29.5539 76.64 29.7375C76.8515 29.3703 77.9087 28.8196 77.9087 29.0032ZM79.1774 27.7183C79.3889 27.7183 78.7545 30.1046 78.5431 29.7375L78.966 28.0854C78.966 28.0854 78.966 27.7183 79.1774 27.7183ZM79.3889 27.1676C79.1774 27.1676 78.966 27.1676 79.1774 26.984C79.3889 26.8004 79.6003 27.1676 79.3889 27.1676ZM81.7148 27.3511C81.9263 27.1676 81.9263 27.1676 81.7148 27.3511C81.7148 27.5347 79.8118 28.6361 79.8118 28.6361C79.6003 28.4525 81.5034 27.3511 81.7148 27.3511ZM81.5034 26.8004C78.5431 28.6361 80.2347 27.1676 81.9263 26.4333C82.1377 26.2497 82.1377 26.2497 81.5034 26.8004ZM82.5606 26.984C83.6179 24.0469 83.8293 24.5976 84.2522 25.699C84.6752 26.6169 84.6751 26.6169 83.8293 27.1676C81.9263 28.4525 82.3492 27.9018 82.5606 26.984ZM83.6179 25.1483C83.195 25.8826 82.7721 26.984 82.5606 27.7183C82.9835 27.5347 83.8293 26.984 84.2522 26.6169C84.0408 26.0662 83.8293 25.5155 83.6179 25.1483ZM87.6355 22.762V22.9456C89.3271 22.2113 89.3271 24.4141 87.6355 25.3319V25.5155L87.2126 25.699V25.5155C85.521 26.2497 85.521 24.0469 87.2126 23.1291V22.9456L87.6355 22.762ZM87.2126 23.3127C86.1553 23.8634 85.9439 25.699 87.2126 25.1483V23.3127ZM87.6355 25.1483C88.6927 24.4141 88.9042 22.762 87.6355 23.3127V25.1483ZM78.966 33.0416C79.1774 32.8581 79.1774 33.0417 78.966 33.2252C78.7545 33.4088 78.3316 33.5923 78.1202 33.7759C78.1202 34.143 77.6973 34.8773 77.9087 34.8773C78.5431 34.5102 79.1774 33.9595 79.1774 33.9595C79.6003 33.7759 78.1202 35.0609 77.6973 35.0609C77.4858 35.0609 77.6973 34.5102 77.6973 34.3266C78.3316 32.491 77.6973 32.491 79.6003 31.7567C79.8118 31.5731 80.0232 31.7567 79.1774 32.1238C78.966 32.3074 78.7545 32.3074 78.7545 32.491C78.5431 32.491 78.5431 32.6745 78.5431 32.6745C78.5431 32.8581 78.3316 33.4088 78.3316 33.5923C78.3316 33.4088 78.7545 33.2252 78.966 33.0416ZM80.4461 31.7567C80.6576 31.7567 80.0232 33.7759 79.8118 33.9595C79.8118 33.9595 79.6003 33.9595 79.8118 33.7759C80.0232 32.6745 80.0232 32.8581 80.2347 32.3074C80.2347 32.1238 80.2347 31.7567 80.4461 31.7567ZM80.6576 31.0224C80.869 31.3896 80.2347 31.3896 80.4461 31.0224C80.6576 30.8389 80.6576 30.8389 80.6576 31.0224ZM83.195 31.206C83.195 31.206 82.5606 31.7567 82.3492 31.7567C82.3492 31.7567 81.0805 32.6745 80.869 32.491C80.869 32.491 83.195 31.0224 83.195 31.206ZM83.195 30.4717L82.7721 30.6553C80.0232 32.6745 81.0805 31.3896 83.195 30.4717C83.4064 30.4717 83.4064 30.2882 83.195 30.4717ZM83.8293 29.3703C85.7324 28.0854 85.9439 29.1868 85.098 29.921C86.1553 29.5539 85.9439 31.206 83.8293 31.9403V29.3703ZM84.2522 30.2882C85.7324 29.5539 85.521 28.6361 84.2522 29.3703V30.2882ZM84.2522 31.5731C84.6752 31.3896 85.3095 31.0224 85.3095 30.4717C85.3095 29.921 84.6752 30.2882 84.2522 30.4717V31.5731ZM85.9439 30.8389C85.9439 30.6553 85.9439 30.8389 85.9439 30.8389C86.5782 28.4525 86.5782 27.7183 86.5782 27.7183C86.5782 27.7183 86.7897 27.7183 86.7897 27.9018C86.7897 28.4525 86.5782 29.3703 86.3668 30.6553C86.1553 30.4717 86.1553 30.8389 85.9439 30.8389ZM88.9042 27.5347C89.3271 27.9018 89.1156 29.3703 88.0584 29.921C87.6355 30.1046 87.2126 30.1046 87.2126 29.5539C87.2126 29.3703 87.2126 28.6361 87.2126 28.4525L87.6355 28.2689C87.6355 28.2689 87.6355 29.1868 87.6355 29.3703C87.6355 29.7375 87.8469 29.7375 88.0584 29.7375C88.6927 29.3703 88.6927 28.2689 88.4813 27.9018L88.9042 27.5347ZM30.9666 61.1272C30.9666 61.1272 30.7551 61.3107 30.7551 61.4943C30.5437 61.6779 30.1208 62.5957 29.9093 62.7793C30.1208 62.9628 30.9666 63.3299 30.7551 63.5135C30.5437 63.6971 29.9093 63.1464 29.9093 62.9628C29.4864 63.6971 29.275 64.0642 29.275 64.2478C29.0635 64.6149 28.8521 64.4313 29.0635 64.4313C29.0635 64.4313 29.275 64.0642 29.9093 62.9628C29.6979 62.7792 29.0635 62.045 29.275 61.8614H29.4864C29.4864 62.045 29.6979 62.2285 29.9093 62.5957C30.1208 62.2285 30.5437 61.4943 30.7551 61.3107C30.7551 60.9436 30.9666 60.9436 30.9666 61.1272ZM31.6009 62.7793C31.6009 62.9628 32.6582 62.2286 32.8696 62.2286C33.504 61.6779 32.8696 62.5957 31.8124 63.1464C31.6009 63.3299 31.3895 63.3299 31.3895 63.1464L32.0238 60.76C32.4467 60.2093 31.6009 62.4121 31.6009 62.7793ZM34.3498 60.76C36.0414 59.8422 35.83 60.0258 35.407 60.3929C32.8696 61.8614 33.2925 61.3107 34.3498 60.76ZM35.83 59.2915L35.407 59.4751C32.2353 61.4943 33.9269 60.0258 35.83 59.2915C35.83 59.1079 35.83 59.1079 35.83 59.2915ZM37.9445 59.6586C38.3674 59.4751 38.5788 58.1901 37.9445 58.0065L38.3674 57.823C38.5788 57.823 38.7903 58.1901 38.7903 58.5572C38.7903 59.8422 37.733 60.3929 37.5216 59.8422C37.5216 60.0258 37.3101 60.3929 36.8872 60.3929C36.0414 60.9436 35.83 59.4751 36.4643 58.7408L36.8872 58.5572C36.4643 59.1079 36.4643 60.3929 36.8872 60.2093C37.3101 60.0258 37.3101 59.2915 37.3101 58.9244L37.733 58.7408C37.733 59.1079 37.5216 59.8422 37.9445 59.6586ZM39.6361 58.7408C40.2704 58.3737 40.9048 57.823 40.9048 57.823C41.1162 57.6394 40.6933 58.0065 40.4819 58.3737C40.059 58.7408 39.2132 59.2915 39.2132 58.9244C39.2132 58.9244 39.8475 56.1709 39.6361 56.538C39.6361 56.538 39.6361 56.3545 39.8475 56.538C40.2704 56.7216 39.2132 59.1079 39.6361 58.7408ZM43.6536 55.9873C43.6536 55.9873 43.0193 56.538 42.8078 56.538C42.8078 56.538 41.5391 57.4559 41.3277 57.2723C41.3277 57.2723 43.6536 55.8038 43.6536 55.9873ZM41.7506 56.538C41.5391 56.538 41.5391 56.538 41.7506 56.538C41.5391 56.3545 43.4422 55.2531 43.6536 55.2531C44.4994 54.7024 42.3849 56.1709 41.7506 56.538ZM44.9223 54.3352L44.7109 54.5188C44.4994 54.5188 45.3453 53.4174 45.7682 53.601C46.614 53.9681 44.9223 55.8038 44.288 56.3545C44.4994 56.1709 45.5567 55.6202 45.7682 55.4366H45.9796C45.9796 55.6202 44.7109 56.3545 44.4994 56.3545H44.288C44.288 56.3545 44.288 56.1709 44.4994 56.1709C44.9223 55.8038 44.9223 55.6202 45.3453 55.2531C46.1911 53.9681 45.7682 53.2338 44.9223 54.3352ZM47.2483 53.4174C45.7682 54.1517 47.4598 52.8667 48.9399 52.316C49.1514 52.1325 49.3628 52.316 48.517 52.6832V52.8667C48.3056 53.2338 48.3056 53.9681 48.517 54.3352C48.7285 54.5188 48.3056 54.7024 48.3056 54.5188C48.0941 54.1517 48.3056 53.2338 48.517 52.8667C48.3056 53.0503 47.8827 53.2338 47.4598 53.4174C47.4598 53.601 47.4598 53.9681 47.4598 54.5188C47.2483 55.8038 46.614 55.8038 47.0369 55.2531C47.2483 54.8859 47.4598 54.3352 47.4598 53.7845C47.2483 53.601 47.2483 53.4174 47.2483 53.4174ZM49.7857 53.7845C50.4201 53.4174 51.0544 52.8667 51.0544 52.8667C51.6888 52.4996 49.3628 54.5188 49.3628 53.7845C49.3628 53.7845 49.9972 51.0311 49.7857 51.3982C49.7857 51.3982 49.7857 51.2146 49.9972 51.3982C50.4201 51.7653 49.3628 53.9681 49.7857 53.7845ZM51.6888 51.2146C52.5346 53.0503 51.9002 51.7653 52.3231 52.4996C52.5346 51.9489 52.9575 51.2146 52.9575 50.6639L53.3804 50.4804C53.3804 51.0311 52.9575 52.1325 52.5346 52.6832L52.1117 52.8667L51.2659 51.3982L51.6888 51.2146ZM54.0147 35.428C54.0147 35.428 54.0147 35.2444 52.9575 37.0801C52.9575 37.0801 53.1689 37.2637 53.3804 37.2637C54.2262 37.8144 53.8033 37.9979 52.746 37.2637C52.5346 37.6308 51.6888 39.0993 51.6888 38.7322C51.6888 38.7322 52.1117 37.9979 52.5346 37.2637C52.1117 36.8965 51.9002 36.5294 51.9002 36.3458C51.9002 36.1623 52.1117 36.1623 52.1117 36.3458C52.1117 36.5294 52.3231 36.713 52.5346 37.0801C52.746 36.713 53.1689 35.9787 53.3804 35.7951C53.8033 35.2444 54.0147 35.428 54.0147 35.428ZM55.4949 35.0609C55.7064 34.8773 55.9178 35.0609 55.9178 35.0609C55.9178 35.2444 55.9178 35.2444 55.7064 35.0609C55.072 34.8773 53.8033 37.8144 55.2835 36.5294L55.7064 36.1623C56.1293 35.9787 55.072 37.0801 54.4376 37.0801C54.2262 37.0801 54.2262 36.8965 54.2262 36.713C54.2262 36.5294 55.072 35.428 55.4949 35.0609ZM58.2438 34.5102C58.4552 34.3266 58.6667 34.5102 58.2438 34.6937C54.6491 36.8965 56.5522 35.428 58.2438 34.5102ZM58.4552 33.7759L58.0323 33.9595C57.1865 34.5102 56.5522 34.8773 56.3407 34.8773H56.1293C56.3407 34.8773 58.2438 33.7759 58.4552 33.7759C58.6667 33.7759 58.6667 33.5923 58.4552 33.7759ZM59.301 34.5102C59.301 33.9595 59.5125 33.2252 59.5125 32.6745C59.5125 32.491 59.7239 32.491 59.7239 32.6745C59.5125 34.8773 59.0896 35.7951 59.301 34.5102ZM60.1468 34.8773C59.9354 34.8773 61.627 30.8389 61.627 31.0224C61.627 30.8389 61.8384 31.0224 61.8384 31.206L60.3583 34.6937C60.3583 34.8773 60.3583 35.0609 60.1468 34.8773ZM63.9529 32.6745C64.3758 32.491 64.5873 31.206 63.9529 31.0224L64.3758 30.8389C64.5873 30.8389 64.7987 31.206 64.7987 31.5731C64.7987 32.8581 63.7415 33.4088 63.53 32.8581C63.53 33.0416 63.3186 33.4088 62.8957 33.4088C62.0499 33.9595 61.8384 32.491 62.4728 31.7567L62.8957 31.5731C62.6842 31.9403 62.4728 32.3074 62.4728 32.6745C62.4728 33.5923 63.3186 33.2252 63.3186 32.491V31.7567L63.7415 31.5731C63.53 32.1238 63.3186 32.8581 63.9529 32.6745ZM67.1247 29.0032C66.9133 29.0032 66.4904 29.3703 66.4904 29.5539C65.0102 31.3896 65.856 32.1238 66.9133 31.0224C67.7591 30.1046 67.1247 31.3896 66.2789 31.7567C64.5873 32.491 65.6446 29.5539 67.1247 29.0032C67.3362 29.0032 67.1247 28.8196 67.1247 29.0032ZM52.3231 40.7514C52.746 40.5678 51.9002 43.5049 51.9002 43.3213C51.6888 43.3213 52.1117 42.2199 52.1117 42.0364C52.3231 40.935 52.3231 40.7514 52.3231 40.7514ZM55.7064 40.5678C55.4949 41.6692 55.072 41.8528 55.2835 41.3021C55.7064 40.2007 55.7064 39.0993 54.8605 40.7514C54.0147 42.0364 53.8033 43.1377 54.2262 41.3021C54.4376 40.5678 54.2262 40.2007 53.8033 41.1185C53.5918 41.4857 52.746 42.9542 52.746 42.9542C52.5346 43.1377 52.9575 41.6692 52.9575 41.4857C52.9575 40.7514 53.3804 40.7514 52.9575 42.5871C53.1689 42.2199 53.8033 40.7514 54.2262 40.3843C54.4376 40.3843 54.6491 40.3843 54.6491 40.7514C54.6491 40.935 54.6491 41.1185 54.4376 41.3021C54.6491 40.7514 54.8605 40.3843 55.072 40.2007C55.7064 39.2829 56.1293 39.2829 55.7064 40.5678ZM58.0323 39.65C57.1865 41.3021 55.7064 40.935 57.398 38.9157C58.2438 37.8144 58.6667 38.365 58.4552 38.9157C58.4552 39.4664 58.2438 39.0993 58.2438 39.8336C58.2438 40.0171 58.2438 40.0171 58.0323 40.0171C58.0323 40.2007 58.0323 39.65 58.0323 39.65ZM58.2438 38.9157C58.2438 38.7322 58.2438 38.7322 58.2438 38.5486C58.0323 38.365 57.6094 38.7322 57.6094 39.0993C55.9178 41.1185 57.398 40.935 58.2438 38.9157ZM60.1468 39.0993C60.1468 39.0993 59.9354 39.0993 59.5125 38.7322C59.301 39.0993 58.6667 40.2007 58.6667 39.8336L59.5125 38.7322C59.301 38.5486 58.8781 38.1815 59.0896 37.9979C59.0896 37.9979 59.301 37.8144 59.301 37.9979C59.301 38.1815 59.5125 38.365 59.7239 38.5486C59.7239 38.365 60.3583 37.2637 60.5697 37.2637C60.9926 37.0801 60.3583 37.8144 60.1468 38.1815C59.9354 38.365 59.9354 38.7322 59.7239 38.7322C60.5697 39.2829 59.301 38.1815 60.1468 39.0993C60.3583 39.0993 60.3583 38.9157 60.1468 39.0993ZM60.5697 39.0993C60.5697 39.0993 62.0499 35.0609 62.0499 35.2444C62.2613 35.428 61.627 36.5294 61.627 36.713C61.2041 37.6308 61.2041 37.9979 60.7812 38.9157C60.7812 39.0993 60.7812 39.2829 60.5697 39.0993ZM65.0102 34.6937C64.5873 35.428 64.5873 35.7951 64.1644 36.5294C64.1644 36.713 63.7415 37.6308 63.53 37.8144H63.3186C63.1071 37.6308 63.1071 37.0801 63.1071 36.713L62.8957 36.8965C62.4728 37.4472 62.4728 37.2637 62.6842 36.8965C62.8957 36.5294 63.1071 36.3458 63.1071 36.713C63.1071 36.8965 63.1071 36.8965 63.1071 37.0801C63.1071 37.2637 63.1071 37.2637 63.1071 37.4472C63.1071 37.6308 63.1071 37.6308 63.3186 37.8144C63.7415 36.8965 63.9529 36.1623 64.3758 35.0609C65.2216 33.9595 65.2216 34.143 65.0102 34.6937ZM65.2216 34.5102L65.0102 34.6937C64.7987 34.6937 65.0102 34.3266 65.2216 34.3266C66.0675 33.4088 66.9133 33.9595 65.856 35.2444C65.4331 35.7951 65.2217 36.1623 64.7987 36.713C64.7987 36.713 65.2216 36.5294 65.4331 36.3458C66.7018 35.6116 66.4904 35.7951 65.6446 36.3458C64.3758 37.0801 64.1644 37.0801 65.0102 36.1623C66.7018 34.143 66.0675 33.5923 65.2216 34.5102ZM66.9133 36.1623C66.9133 35.9787 67.3362 35.6116 67.1247 35.6116H66.9133C66.9133 35.428 67.3362 35.2444 67.3362 35.6116C67.3362 35.7951 67.3362 35.9787 67.1247 36.1623C66.7018 36.5294 66.4904 36.713 66.9133 36.1623ZM59.0896 49.379C59.0896 49.5625 57.6094 50.4804 57.398 50.4804C56.7636 50.2968 59.0896 48.2776 59.0896 47.7269L58.4552 47.9104C58.2438 48.094 57.6094 48.4611 57.6094 48.2776C57.6094 47.9104 59.9354 46.8091 59.301 48.094C59.301 48.2776 57.6094 50.2968 57.6094 50.2968C57.6094 50.4804 59.0896 49.1954 59.0896 49.379ZM60.3583 47.9104C63.1071 46.2584 61.2041 47.7269 59.5125 48.6447C59.5125 48.4611 60.3583 48.094 60.3583 47.9104ZM61.8384 46.6255L61.4155 46.8091C58.4552 48.6447 59.7239 47.5433 61.8384 46.6255C61.8384 46.4419 61.8384 46.4419 61.8384 46.6255ZM64.7987 44.7898C64.1644 46.0748 63.53 47.5433 63.53 47.5433C63.53 47.7269 63.3186 47.9104 63.1071 47.9104C62.8957 47.9104 62.6842 46.9926 62.6842 46.8091C62.4728 46.9926 62.4728 47.1762 62.4728 47.1762C62.2613 47.5433 62.2613 46.8091 62.8957 46.6255C63.1071 46.4419 63.1071 46.8091 63.1071 46.9926C63.1071 47.1762 63.1071 47.5433 63.3186 47.7269C63.7415 46.8091 63.9529 46.0748 64.3758 44.9734C64.7987 44.2391 65.0102 44.2391 64.7987 44.7898ZM65.0102 46.6255H65.6446C65.6446 46.8091 65.0102 46.9926 64.5873 46.8091C63.7415 46.4419 64.3758 44.9734 64.7987 44.2391C65.856 42.4035 66.2789 42.5871 66.0675 42.9542C65.6446 43.3213 65.4331 43.6884 65.0102 44.0556C64.5873 45.157 63.9529 46.6255 65.0102 46.6255ZM68.3934 42.5871C68.3934 42.5871 68.3934 42.4035 67.3362 44.2391C67.3362 44.2391 67.5476 44.4227 67.7591 44.4227C68.6049 44.9734 68.182 45.157 67.1247 44.4227C66.9133 44.7898 66.0675 46.2584 66.0675 45.8912C66.0675 45.8912 66.4904 45.157 66.9133 44.4227C66.7018 44.0556 66.4904 43.872 66.4904 43.5049C66.4904 43.3213 66.7018 43.3213 66.7018 43.5049C66.7018 43.6884 66.9133 43.872 67.1247 44.2391C67.3362 43.872 67.7591 43.1377 67.9705 42.9542C68.3934 42.4035 68.6049 42.5871 68.3934 42.5871ZM70.085 42.2199C70.5079 42.0364 70.7194 42.4035 70.2965 42.4035C69.6621 42.2199 68.182 45.5241 70.085 43.6884C70.2965 43.5049 70.5079 43.5049 70.085 43.872C69.6621 44.2391 69.0278 44.6063 69.0278 44.0556C68.8163 43.6884 69.6621 42.5871 70.085 42.2199ZM70.9308 42.7706C70.7194 42.7706 72.1995 41.6692 72.1995 41.8528C72.1995 42.0364 70.9308 42.7706 70.9308 42.7706ZM74.7369 39.4664C74.7369 39.4664 74.5255 39.65 74.5255 39.8336C74.314 40.0171 73.8911 40.935 73.6797 41.1185C73.6797 41.1185 73.8911 41.3021 74.1026 41.3021C74.9484 41.8528 74.5255 41.8528 74.1026 41.6692C73.8911 41.6692 73.6797 41.4857 73.4682 41.3021C72.411 43.1377 72.411 42.7706 72.411 42.5871C72.6224 42.2199 73.0453 41.4857 73.2568 41.1185C73.0453 40.935 72.8339 40.7514 72.8339 40.3843C72.8339 40.2007 73.0453 40.2007 73.0453 40.3843C73.0453 40.5678 73.2568 40.7514 73.4682 41.1185C73.6797 40.7514 74.1026 40.0171 74.314 39.8336C74.7369 39.2829 74.9484 39.4664 74.7369 39.4664ZM75.5827 41.1185C76.2171 40.7514 76.8515 40.2007 76.8515 40.2007C77.4858 39.8336 75.1598 41.8528 75.1598 41.1185C75.1598 41.1185 75.7942 38.365 75.5827 38.7322C75.5827 38.7322 75.5828 38.5486 75.7942 38.7322C76.2171 39.0993 75.1598 41.4857 75.5827 41.1185ZM77.0629 40.935C77.0629 40.935 78.3316 39.4664 78.5431 38.5486C78.7545 38.1815 78.966 37.2637 78.5431 37.0801C78.5431 37.0801 78.3316 37.0801 78.1202 36.8965C77.9087 36.8965 78.1202 36.713 78.1202 36.713C78.7545 36.3458 79.6003 37.4472 78.1202 39.65C77.2744 40.935 77.0629 40.935 77.0629 40.935ZM80.4461 36.8965L80.2347 37.0801C80.0232 37.0801 80.2347 36.713 80.4461 36.713C81.2919 35.7951 82.3492 36.3458 80.6576 38.365C80.4461 38.5486 80.2347 38.9157 80.0232 39.0993C80.2347 39.0993 81.5034 37.9979 81.5034 38.1815C81.5034 38.365 80.2347 39.0993 80.0232 39.0993C79.8118 39.0993 79.8118 39.0993 79.8118 39.0993C79.8118 38.9157 80.2347 38.5486 80.4461 38.365C81.9263 36.5294 81.0805 35.9787 80.4461 36.8965ZM83.4064 36.3458C83.4064 36.5294 83.4064 36.713 83.195 36.8965C83.195 36.8965 83.195 37.2637 82.9835 37.2637C82.9835 37.2637 82.9835 37.2637 82.9835 37.0801C82.9835 37.2637 82.9835 37.0801 83.4064 36.3458C83.195 36.5294 82.3492 36.8965 82.3492 36.8965C82.3492 36.8965 82.1377 36.713 82.3492 36.713C83.4064 36.1623 83.195 36.3458 83.4064 35.428C83.4064 35.2444 83.4064 35.2444 83.6179 35.2444C83.8293 35.2444 83.6179 35.7951 83.6179 35.9787C83.8293 35.7951 84.4637 35.428 84.6751 35.428C84.4637 35.428 84.6751 35.6116 83.4064 36.3458ZM84.6751 36.713C84.4637 36.713 84.8866 35.7951 84.8866 35.6116C85.098 35.0609 85.098 34.6937 85.098 34.5102C84.6751 34.5102 85.7324 33.7759 86.3668 33.7759C86.7897 33.7759 86.5782 34.3266 86.3668 34.5102C86.1553 34.8773 85.9439 35.0609 85.521 35.428C85.3095 35.6116 85.098 35.6116 85.9439 35.7951C85.9439 35.7951 86.1553 35.7951 86.3668 35.7951C86.5782 35.7951 86.5782 35.9787 86.3668 35.9787C86.1553 36.1623 85.3095 35.7951 85.098 35.6116C85.098 35.7951 85.098 35.7951 85.098 35.9787C85.098 36.1623 85.098 36.3458 84.8866 36.5294C84.8866 36.5294 84.8866 36.713 84.6751 36.713ZM85.3095 35.428C85.521 35.0609 86.5782 34.5102 86.5782 33.9595C86.5782 33.5923 85.9439 33.9595 85.521 34.3266C85.521 34.5102 85.3095 35.2444 85.3095 35.428ZM87.8469 33.0416L87.6355 33.2252C87.424 33.2252 88.2698 32.1238 88.6927 32.3074C88.9042 32.491 88.9042 32.8581 88.6927 33.2252C88.4813 33.5923 88.0584 34.143 87.8469 34.5102C87.6355 34.6937 87.424 35.0609 87.2126 35.2444C87.424 35.0609 88.4813 34.5102 88.6927 34.3266H88.9042C88.9042 34.5102 87.6355 35.428 87.2126 35.428C87.0011 35.428 87.424 34.8773 87.6355 34.6937C88.0584 34.3266 88.2698 33.9595 88.4813 33.5923C88.6927 33.2252 88.6927 32.6745 88.2698 32.8581C88.2698 32.8581 88.0584 33.0416 87.8469 33.0416Z"
-          fill="white"
-        />
-        <path
-          d="M48.5169 98.2076L28.6405 88.2951C28.6405 88.1115 28.4291 88.1115 28.4291 87.7444L48.3055 97.6569C48.3055 98.024 48.5169 98.2076 48.5169 98.2076Z"
-          fill="#B09529"
-        />
-        <path
-          d="M48.3055 97.8404L28.4291 87.7443V85.5415L48.3055 95.454V97.8404Z"
-          fill="#D3AB34"
-        />
-        <path
-          d="M49.1517 94.3528L29.2753 84.4403L43.654 77.0977L63.5304 87.1938L49.1517 94.3528Z"
-          fill="#FFCD39"
-        />
-        <path
-          d="M48.3055 95.4541L28.4291 85.5416C28.4291 85.1744 28.852 84.6238 29.2749 84.4402L49.1513 94.3527C48.7284 94.5363 48.3055 95.087 48.3055 95.4541Z"
-          fill="#EEC23E"
-        />
-        <path
-          d="M49.1517 98.2075C48.7288 98.3911 48.3059 98.2075 48.3059 97.8404V95.4541C48.3059 95.0869 48.7288 94.5362 49.1517 94.3527L63.5304 87.1936V87.5607L63.1075 87.7443L62.2617 89.2128L63.1075 90.4978L63.5304 90.6814V90.8649L49.1517 98.2075Z"
-          fill="#B09529"
-        />
-        <path
-          d="M48.5174 97.8407C48.5174 98.0243 48.5174 98.0243 48.5174 98.0243H48.7288L63.1075 90.8652L62.6846 90.6817L61.8388 89.5803L62.6846 87.7446L48.9403 94.7201C48.7288 94.7201 48.5174 94.9037 48.5174 95.0872C48.3059 95.2708 48.3059 95.4544 48.3059 95.6379L48.5174 97.8407Z"
-          fill="#B09529"
-        />
-        <path
-          d="M63.1074 87.561L49.3631 94.5365C49.1516 94.5365 48.9402 94.7201 48.9402 94.9037C48.7287 95.0872 48.7287 95.2708 48.7287 95.4543V97.6571L63.1074 90.4981V87.561Z"
-          fill="#FFFFFA"
-        />
-        <path
-          d="M38.1565 83.8898L43.0199 81.5034L42.597 81.3198L37.7336 83.7062L38.1565 83.8898Z"
-          fill="#60522B"
-        />
-        <path
-          d="M37.945 85.1747L45.7687 81.3198L45.3458 81.1362L37.5221 84.9911L37.945 85.1747Z"
-          fill="#60522B"
-        />
-        <path
-          d="M39.214 85.909L46.8262 82.0542L46.4033 81.8706L38.7911 85.7255L39.214 85.909Z"
-          fill="#60522B"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Sittingperson;
