+++ src/app/(dashboard)/admin/dashboard/account/page.tsx
@@ -1,0 +1,5 @@
+import AccountProfile from "@/app/(dashboard)/components/AccountProfile";
+
+export default function Page() {
+  return <AccountProfile />;
+}
