+++ src/app/(dashboard)/admin/dashboard/transactions/page.tsx
@@ -1,0 +1,16 @@
+"use client";
+import React, { useId, useState } from "react";
+import DashHeader from "@/app/(dashboard)/components/DashHeader";
+
+
+
+const Page = () => {
+
+  return (
+    <div className="">
+      <DashHeader title="TRANSACTIONS" />
+    </div>
+  );
+};
+
+export default Page;
