+++ src/app/(dashboard)/client/dashboard/exams/page.tsx
@@ -1,0 +1,72 @@
+import { Clock, Flag } from "lucide-react";
+import Link from "next/link";
+import React from "react";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableFooter,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+import { quizhistory } from "@/lib/utils";
+import StartExamCard from "@/app/(dashboard)/components/StartExamCard";
+import DashHeader from "@/app/(dashboard)/components/DashHeader";
+
+const page = () => {
+  return (
+    <>
+      <div className="flex min-h-screen flex-col gap-4">
+        <DashHeader title="EXAMS" />
+        <div className="grid grid-cols-1 gap-4 md:grid-cols-1 lg:grid-cols-3">
+          <StartExamCard />
+          <div className="col-span-2 flex min-h-60 flex-row items-end gap-4 rounded-2xl border-2 border-black/10 bg-white p-8 shadow-2xl">
+            <div>
+              <h1 className="text-[8rem]">10</h1>
+              <p>Imyitozo wakoze </p>
+            </div>
+            <div className="flex min-w-64 flex-col gap-2 border-l-4 border-black pl-8">
+              <h6 className="inline-flex justify-between">
+                imyitozo watsinze: <span>05</span>
+              </h6>
+              <h6 className="inline-flex justify-between">
+                imyitozo watsinzwe: <span>05</span>
+              </h6>
+              <h6 className="inline-flex justify-between">
+                Imyitozo itararangira: <span>05</span>
+              </h6>
+            </div>
+          </div>
+        </div>
+        <div className="p-4">
+          <h1>Imyitozo iheruka</h1>
+          <div className="min-h-72 w-full rounded-2xl p-4 shadow-2xl">
+            <Table>
+              <TableHeader>
+                <TableRow>
+                  <TableHead className="p-4">Itariki</TableHead>
+                  <TableHead className="p-4">Amanota</TableHead>
+                  <TableHead className="p-4">Iminota wakoresheje</TableHead>
+                  <TableHead className="p-4">Ikigero</TableHead>
+                </TableRow>
+              </TableHeader>
+              <TableBody>
+                {quizhistory.map((item, index) => (
+                  <TableRow key={index}>
+                    <TableCell className="p-4">{item.date}</TableCell>
+                    <TableCell className="p-4">{item.marks}</TableCell>
+                    <TableCell className="p-4">{item.time}</TableCell>
+                    <TableCell className="p-4">{item.status}</TableCell>
+                  </TableRow>
+                ))}
+              </TableBody>
+            </Table>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default page;
