+++ src/app/(dashboard)/client/dashboard/layout.tsx
@@ -1,0 +1,7 @@
+import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
+import { AppSidebar } from "@/app/(dashboard)/components/AppSidebar";
+import Navbar from "../../components/Navbar";
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return <section className="w-full">{children}</section>;
+}
