+++ src/app/(dashboard)/client/dashboard/payment/page.tsx
@@ -1,0 +1,12 @@
+import DashHeader from '@/app/(dashboard)/components/DashHeader'
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>
+      <DashHeader title="PAYMENT" />
+    </div>
+  )
+}
+
+export default page
